% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceRelations.R
\name{actualLoadRelationTypes}
\alias{actualLoadRelationTypes}
\title{API request to retrieve all registered relation types}
\usage{
actualLoadRelationTypes(conn, ...)
}
\arguments{
\item{conn}{database connection}

\item{...}{args}
}
\description{
API request to retrieve all registered relation types
}
