% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceRelations.R
\name{actualLoadResourceRelations}
\alias{actualLoadResourceRelations}
\title{API reqeust to retrieve all registered resource relations}
\usage{
actualLoadResourceRelations(resourceId)
}
\arguments{
\item{resourceId}{resource}
}
\description{
API reqeust to retrieve all registered resource relations
}
\references{
ics1044
}
