% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{actualLoadReviews}
\alias{actualLoadReviews}
\title{API reqeust to retrieve all registered reviews}
\usage{
actualLoadReviews(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
API reqeust to retrieve all registered reviews
}
\references{
ics348
}
