% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{addExtLink}
\alias{addExtLink}
\title{adds a new external link to the step}
\usage{
addExtLink(stepHandle, name, url)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{name}{name of the link}

\item{url, }{the url the link targets}
}
\description{
adds a new external link to the step
}
