% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_realise.R
\name{addExtLinkToStep_deprecated}
\alias{addExtLinkToStep_deprecated}
\title{Add external link to step - deprecated version}
\usage{
addExtLinkToStep_deprecated(newStep, linkPrep)
}
\description{
Add external link to step - deprecated version
}
\keyword{internal}
