% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_realise.R
\name{addFileToStep_deprecated}
\alias{addFileToStep_deprecated}
\title{Add file to step - deprecated version}
\usage{
addFileToStep_deprecated(newStep, filePrep, isLocal, env)
}
\description{
Add file to step - deprecated version
}
\keyword{internal}
