% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTreeConnections.R
\name{addMultiTreeLineage}
\alias{addMultiTreeLineage}
\title{addMultiTreeLineage
internal recursion function for getFullLineage}
\usage{
addMultiTreeLineage(
  workflowHandle,
  stepHandle,
  workflowEnv = new.env(),
  depth = -1
)
}
\arguments{
\item{workflowHandle}{the retrieved workflow}

\item{stepHandle}{the currently investigated step}

\item{workflowEnv}{internal variable}

\item{depth, }{the depth of recursion when retrieving additional trees. -1 takes all trees into account}
}
\description{
addMultiTreeLineage
internal recursion function for getFullLineage
}
