% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{addStepLineage}
\alias{addStepLineage}
\title{adds a new lineage and implicitly usage to the step}
\usage{
addStepLineage(stepHandle, lineageHandle)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{lineageHandle}{the ident of the file in the repository}
}
\description{
adds a new lineage and implicitly usage to the step
}
