% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{addStepLocalFile}
\alias{addStepLocalFile}
\title{adds a new local file to the step}
\usage{
addStepLocalFile(stepHandle, path, name = NULL, variableName = NULL)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{path}{local path to the file}

\item{name}{leave empty if you want to use the same name as the used file}

\item{variableName}{the name of the variable the file should be bound to, optional}
}
\description{
adds a new local file to the step
}
