% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportImportUtils.R
\name{applyOutsideLinkMapping}
\alias{applyOutsideLinkMapping}
\title{Apply outside link mapping to step templates}
\usage{
applyOutsideLinkMapping(importWF, workflowTemplate, linkMapping)
}
\arguments{
\item{importWF}{Data frame from workflow.json.}

\item{workflowTemplate}{Workflow template environment.}

\item{linkMapping}{Environment with version-to-entityId mappings.}
}
\description{
Apply outside link mapping to step templates
}
\keyword{internal}
