% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenManagement.R
\name{applyTokenData}
\alias{applyTokenData}
\title{Apply token data to session}
\usage{
applyTokenData(tokenData)
}
\arguments{
\item{tokenData}{List containing token information with fields:
IMPROVER_TOKEN, IMPROVER_TOKEN_EXPIRATION, IMPROVER_REFRESH_TOKEN,
IMPROVER_LAST_ACCESS, IMPROVER_REPO_URL, IMPROVER_USER}
}
\description{
Applies a complete token data structure to the current session
}
