% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportImportUtils.R
\name{applyToolMappingFromFile}
\alias{applyToolMappingFromFile}
\title{Apply tool mapping from file to step templates}
\usage{
applyToolMappingFromFile(zipFile, baseName, workflowTemplate)
}
\arguments{
\item{zipFile}{Path to original zip file.}

\item{baseName}{Base name for tool mapping file lookup.}

\item{workflowTemplate}{Workflow template environment.}
}
\description{
Apply tool mapping from file to step templates
}
\keyword{internal}
