% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticatedREST.R
\name{authenticatetREST}
\alias{authenticatetREST}
\title{authenticatedREST uses the connection information from #ics1081 improveConnect# to create a REST call to the repository
it needs the following parameters
Those values are example values}
\usage{
authenticatetREST(
  url,
  urlParams = list(),
  queryParams = list(),
  data = "",
  restType = "GET",
  contentType = "application/json",
  encode = "json",
  ignoreFail = T
)
}
\arguments{
\item{url}{in the format /resources/{resourceId}}

\item{urlParams}{in the format list(resourceId="1B1D3B817F424BA594893A5013DBFEEA")}

\item{queryParams}{in the format list(isResourceVersion="true")}

\item{data}{this block is written to the body, either list, binary or json}

\item{restType}{character, default is GET, possible values are POST,GET,PUT and DELETE}

\item{contentType}{is directly set to the header, default is application/json, used to specify the format for data}

\item{encode}{is directly set to the header, default is json, should be fine for most rest calls, for file uploads use NULL}

\item{ignoreFail}{is by default set to TRUE, if false, an error is thrown if the error code is not between 200 and 300

in this example the following URL would be constructed:
baseURL from improveConnect /resources/1B1D3B817F424BA594893A5013DBFEEA?isResourceVersion=true

the unparsed result is returned from this call}
}
\description{
authenticatedREST uses the connection information from #ics1081 improveConnect# to create a REST call to the repository
it needs the following parameters
Those values are example values
}
\references{
ics1082
}
