% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshRunner.R
\name{autoRefreshStart}
\alias{autoRefreshStart}
\title{autoRefreshStart}
\usage{
autoRefreshStart(verbose = F)
}
\arguments{
\item{verbose}{if true a log is written to stdout.txt and stderr.txt}
}
\description{
starts a background process to automatically refresh the token
}
