% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{buildCaption}
\alias{buildCaption}
\title{general function to build the caption for a resource}
\usage{
buildCaption(resource)
}
\arguments{
\item{resource}{resource to be described}
}
\description{
general function to build the caption for a resource
}
