% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{buildDescriptor}
\alias{buildDescriptor}
\title{general function to build the descriptor}
\usage{
buildDescriptor(resource, localPath, caption = "", description = "")
}
\arguments{
\item{resource}{resource to be described}

\item{localPath}{path to the file locally}

\item{caption}{by default entityID and lastmodified are the caption, here alternative text can be provided}

\item{description}{by default filename is the caption, here alternative text can be provided}
}
\description{
general function to build the descriptor
}
