% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{byNotEmptyAsDf}
\alias{byNotEmptyAsDf}
\title{byNotEmptyAsDf by that automatically uses the complete dataframe and returns null if an empty or non existing data frame is handed over
it automatically tries to convert the result list to a data frame again}
\usage{
byNotEmptyAsDf(df, func)
}
\arguments{
\item{df}{the data frame}

\item{func}{the function that is applied to each row of the data frame}
}
\description{
byNotEmptyAsDf by that automatically uses the complete dataframe and returns null if an empty or non existing data frame is handed over
it automatically tries to convert the result list to a data frame again
}
