% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_realise.R
\name{changeStepDescription_deprecated}
\alias{changeStepDescription_deprecated}
\title{Change step description using PUT request}
\usage{
changeStepDescription_deprecated(step, description)
}
\description{
Change step description using PUT request
}
\keyword{internal}
