% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_realise.R
\name{changeStepRationale_deprecated}
\alias{changeStepRationale_deprecated}
\title{Change step rationale using PUT request}
\usage{
changeStepRationale_deprecated(step, rationale)
}
\description{
Change step rationale using PUT request
}
\keyword{internal}
