% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{changeStepRemoteFile}
\alias{changeStepRemoteFile}
\title{changes a remote file in the step}
\usage{
changeStepRemoteFile(
  stepHandle,
  name,
  asLink = NULL,
  newName = NULL,
  newIdent = NULL
)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{name}{the name of the file in the inventory}

\item{asLink}{if the file is added as link, boolean, not changed if NULL}

\item{newName}{the new name of the file, not changed if NULL}

\item{newIdent}{the new ident of the file, not changed if NULL}
}
\description{
changes a remote file in the step
}
