% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveConnect.R
\name{clearConnectionData}
\alias{clearConnectionData}
\title{Clear Connection Data}
\usage{
clearConnectionData(includeRepoData = F)
}
\arguments{
\item{includeRepoData}{includes the repository URL and the selected step in the clear process}
}
\description{
deletes repoUrl, stepId, and token from the environment variables and calls improveDisconnect
}
\seealso{
\code{\link[=improveDisconnect]{improveDisconnect()}}
}
