% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliSetup.R
\docType{data}
\name{cliEnv}
\alias{cliEnv}
\title{stores all cli related variables during the load time of the package}
\format{
An object of class \code{environment} of length 0.
}
\usage{
cliEnv
}
\description{
stores all cli related variables during the load time of the package
}
\keyword{datasets}
