% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliSetup.R
\name{cliPath}
\alias{cliPath}
\title{Get CLI Path}
\usage{
cliPath(unpack = TRUE)
}
\arguments{
\item{unpack}{Logical. Whether to unpack the CLI files if not found.}
}
\value{
The path to the CLI executable.
}
\description{
Determines the path to the CLI executable, unpacking it if necessary.
}
