% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliCloneChildstep.R
\name{cloneChildstepCli}
\alias{cloneChildstepCli}
\title{Clone Child Step CLI}
\usage{
cloneChildstepCli(
  localSource,
  localTarget,
  comment,
  toolCategory = NULL,
  tool = NULL
)
}
\arguments{
\item{localSource}{The local source repository.}

\item{localTarget}{The local target repository.}

\item{comment}{A comment for the clone operation.}

\item{toolCategory}{The tool category. Default is NULL.}

\item{tool}{The tool name. Default is NULL.}
}
\description{
Clones a child step from the CLI.
}
