% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportImportUtils.R
\name{collectInputFiles}
\alias{collectInputFiles}
\title{Collect input files (non-link) from workflow step data}
\usage{
collectInputFiles(workflowDf)
}
\arguments{
\item{workflowDf}{Data frame of workflow steps.}
}
\value{
Data frame of input files, or NULL if none found.
}
\description{
Collect input files (non-link) from workflow step data
}
\keyword{internal}
