% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliConfigureUserProfile.R, R/cliSetup.R
\name{configureUserProfile}
\alias{configureUserProfile}
\title{Configure User Profile}
\usage{
configureUserProfile(userProfile = "improveR")

configureUserProfile(userProfile = "improveR")
}
\arguments{
\item{userProfile}{A string representing the user profile name. Default is "improveR".}
}
\description{
Configures the user profile for CLI usage.

Configures the user profile for CLI usage.
}
