% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportImportUtils.R
\name{convertParentIdentsToFullNames}
\alias{convertParentIdentsToFullNames}
\title{Convert parentIdent resourceIds to fullNames for portability}
\usage{
convertParentIdentsToFullNames(workflowDf)
}
\arguments{
\item{workflowDf}{Data frame of workflow steps.}
}
\value{
Modified workflowDf with parentFullName column and parentIdent removed.
}
\description{
Convert parentIdent resourceIds to fullNames for portability
}
\keyword{internal}
