% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenRefresherPlugin.R
\name{createChannelTokenRefresher}
\alias{createChannelTokenRefresher}
\title{Create channel-based token refresher}
\usage{
createChannelTokenRefresher()
}
\value{
Token refresher object
}
\description{
Creates a channel-based token refresher using encrypted channels
}
