% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliCreateContentCache.R
\name{createContentCache}
\alias{createContentCache}
\title{Create Content Cache}
\usage{
createContentCache(localPath)
}
\arguments{
\item{localPath}{The local path for the content cache.}
}
\description{
Creates a content cache for the specified location.
}
