% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createExternalLink}
\alias{createExternalLink}
\title{creates an external link in the repository and returns the resource object}
\usage{
createExternalLink(
  targetIdent,
  linkName = "",
  url,
  comment = "Created by improveRW"
)
}
\arguments{
\item{targetIdent}{one or many target folders, as resource, id or path}

\item{linkName}{one or many linkNames}

\item{url}{url for the external link}

\item{comment}{defaults to Created by improveRW}
}
\description{
creates an external link in the repository and returns the resource object
}
\references{
ics1104
}
