% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createFile}
\alias{createFile}
\title{creates a file in the repository and returns the resource object}
\usage{
createFile(
  targetIdent,
  fileName = "",
  localPath = "",
  comment = "Created by improveRW"
)
}
\arguments{
\item{targetIdent}{one or many target folders, as resource, id or path}

\item{fileName}{one or many fileNames}

\item{localPath}{path to a local file for initial file contents}

\item{comment}{defaults to Created by improveRW}
}
\description{
creates a file in the repository and returns the resource object
}
\references{
ics1102
}
