% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenRefresherPlugin.R
\name{createFileTokenRefresher}
\alias{createFileTokenRefresher}
\title{Create file-based token refresher}
\usage{
createFileTokenRefresher()
}
\value{
Token refresher object
}
\description{
Creates a file-based token refresher using XOR encryption
}
