% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareStep.R
\name{createPreparedStep}
\alias{createPreparedStep}
\title{creates the prepared step in the repository}
\usage{
createPreparedStep(stepHandle)
}
\arguments{
\item{stepHandle}{id of the prepared step}
}
\description{
creates the prepared step in the repository
}
\references{
ics1140
}
