% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_realise.R
\name{createPreparedStep_deprecated}
\alias{createPreparedStep_deprecated}
\title{Create prepared step using old API approach}
\usage{
createPreparedStep_deprecated(env, prepStep)
}
\description{
Create prepared step using old API approach
}
\keyword{internal}
