% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{createProcessFileVariable}
\alias{createProcessFileVariable}
\title{createProcessFileVariable}
\usage{
createProcessFileVariable(ident, processId, name, variableType, position)
}
\arguments{
\item{ident}{id of the step the variables refer to}

\item{processId}{id of the process the variables refer to}

\item{name}{name of the variable}

\item{variableType}{fileRef or filePath}

\item{position}{position of the variable}
}
\description{
createProcessFileVariable
}
