% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceRelations.R
\name{createResourceRelation}
\alias{createResourceRelation}
\title{Creates a New Resource Relation}
\usage{
createResourceRelation(
  resourceId,
  targetResourceId,
  relationTypeId,
  description
)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}

\item{targetResourceId}{id (UUID) of the target resource}

\item{relationTypeId}{id (UUID) of the relation type}

\item{description}{description of the resource relation}
}
\description{
Creates a New Resource Relation
}
\references{
ics1044
}
