% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewEntries.R
\name{createReviewEntry}
\alias{createReviewEntry}
\title{Creates Review Entries For A Review}
\usage{
createReviewEntry(reviewId, resourceIds)
}
\arguments{
\item{reviewId}{id (UUID) of the review}

\item{resourceIds}{vector of resourceIds}
}
\description{
Creates Review Entries For A Review
}
\references{
ics1541
}
