% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewEntryComments.R
\name{createReviewEntryComment}
\alias{createReviewEntryComment}
\title{creates a comment for a review entry}
\usage{
createReviewEntryComment(reviewId, entryId, comment)
}
\arguments{
\item{reviewId}{id (UUID) of the review}

\item{entryId}{id (UUID) of the resource}

\item{comment}{comment}
}
\description{
creates a comment for a review entry
}
\references{
ics1544
}
