% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{createStep}
\alias{createStep}
\title{createStep}
\usage{
createStep(treeIdent, parentStepIdent = NULL, toolId = NULL)
}
\arguments{
\item{treeIdent}{id of the containing tree}

\item{parentStepIdent}{optional id of the parent step}

\item{toolId}{optional id of the tool}
}
\description{
createStep
}
\references{
ics1140
}
