% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGridArguments.R
\name{deleteGridArgumentsById}
\alias{deleteGridArgumentsById}
\title{deleteGridArgumentsById}
\usage{
deleteGridArgumentsById(processId, ids)
}
\arguments{
\item{processId}{id of the process}

\item{ids}{ids of the grid arguments}
}
\description{
deleteGridArgumentsById
}
\references{
ics1222
}
