% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGridArguments.R
\name{deleteGridArgumentsByName}
\alias{deleteGridArgumentsByName}
\title{deleteGridArgumentsByName}
\usage{
deleteGridArgumentsByName(processId, argumentName)
}
\arguments{
\item{processId}{id of the process}

\item{argumentName}{name of the grid argument}
}
\description{
deleteGridArgumentsByName
}
\references{
ics1222
}
