% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceRelations.R
\name{deleteRelationType}
\alias{deleteRelationType}
\title{Delete a Relation Type}
\usage{
deleteRelationType(relationTypeId)
}
\arguments{
\item{relationTypeId}{Character. UUID of the relation type to delete.}
}
\value{
\code{TRUE} if deleted successfully, \code{FALSE} otherwise.
}
\description{
Deletes a relation type from the server and refreshes the cached list.
}
\references{
ics1700
}
