% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewEntries.R
\name{deleteReviewEntry}
\alias{deleteReviewEntry}
\title{Deletes All Review Entries From A Review}
\usage{
deleteReviewEntry(reviewId)
}
\arguments{
\item{reviewId}{id (UUID) of the review}
}
\description{
Deletes All Review Entries From A Review
}
\references{
ics365
}
