% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{detachStep}
\alias{detachStep}
\title{detaches a step from its parent}
\usage{
detachStep(ident, from = pwd())
}
\arguments{
\item{ident}{ident of the step}

\item{from}{from if a relative path is used}
}
\description{
detaches a step from its parent
}
\references{
ics1225
}
