% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{detachWorkflowFromResources}
\alias{detachWorkflowFromResources}
\title{detachWorkflowFromResources
deletes all entity ids from a workflow, switches from in place execution to creation of new steps}
\usage{
detachWorkflowFromResources(workflowHandle)
}
\arguments{
\item{workflowHandle}{handle of the workflow}
}
\description{
detachWorkflowFromResources
deletes all entity ids from a workflow, switches from in place execution to creation of new steps
}
