% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{detachWorkflowFromTrees}
\alias{detachWorkflowFromTrees}
\title{detachWorkflowFromTrees
removes coupling to a specific tree for an entire workflow.
if no treeIdent is provided treeName and treePath have to be provided.}
\usage{
detachWorkflowFromTrees(workflowHandle)
}
\arguments{
\item{workflowHandle}{handle of the workflow}
}
\description{
detachWorkflowFromTrees
removes coupling to a specific tree for an entire workflow.
if no treeIdent is provided treeName and treePath have to be provided.
}
