% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliExecute.R, R/cliSetup.R
\name{executeCli}
\alias{executeCli}
\title{Execute CLI Command}
\usage{
executeCli(cliString)

executeCli(cliString)
}
\arguments{
\item{cliString}{A string representing the CLI command to execute.}
}
\description{
Executes a CLI command using the appropriate CLI executable.

Executes a CLI command using the appropriate CLI executable.
}
