% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRun.R
\name{executeFolder}
\alias{executeFolder}
\title{executeFolder}
\usage{
executeFolder(path)
}
\arguments{
\item{path}{path to the executable package}
}
\description{
Executes all Rmd-files in the folder specified by path. If available,
execution of the Rmd-files is preceeded by the execution of an init.R file and/or
followed by the execution of an cleanup.R file.
}
