% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepWorkflowCopy.R
\name{executeWorkflow}
\alias{executeWorkflow}
\title{executeWorkflow}
\usage{
executeWorkflow(workflow)
}
\arguments{
\item{workflow}{a workflow handle or a workflow data.frame, changes to steps in workflow data.frame will be persisted by execution. consider creating a deep copy}
}
\description{
executeWorkflow
}
\references{
ics1212
}
