% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worflowOrder.R
\name{executionOrder}
\alias{executionOrder}
\title{creates a executable order following the dependencies
One workflow can span multiple analsis trees}
\usage{
executionOrder(workflow)
}
\arguments{
\item{workflow}{the workflow to get ordered}
}
\description{
creates a executable order following the dependencies
One workflow can span multiple analsis trees
}
\references{
ics1231
}
