% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepEquality.R
\name{existsInTargetTree}
\alias{existsInTargetTree}
\title{existsInTargetTree}
\usage{
existsInTargetTree(handle)
}
\arguments{
\item{handle}{check if a step with the same configuration already exists in the tree}
}
\description{
existsInTargetTree
}
\references{
ics1228
}
