% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliSetup.R
\name{filterOsVersion}
\alias{filterOsVersion}
\title{Filter OS Version}
\usage{
filterOsVersion(osKey, possibleFolders)
}
\arguments{
\item{osKey}{A string representing the OS key (e.g., "win32", "lnx", "macosx").}

\item{possibleFolders}{a list of folders that could contain the improve-cli script}
}
\value{
The path to the script if found, otherwise NULL.
}
\description{
Filters the possible folders to find the one matching the given OS key.
}
