% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStep.R
\name{finishRun}
\alias{finishRun}
\title{finishRun}
\usage{
finishRun(handle, runserverName = NULL, runserverToolName = NULL)
}
\arguments{
\item{handle}{check if a step with the same configuration already exists in the tree}

\item{runserverName}{if this is set, the step only counts as finished if it was finished with this runserver (needs to be combined with tool), overridden by settings in stephandle}

\item{runserverToolName}{if this is set, the step only counts as finished if it was finished with this tool (needs to be combined with runserver), overridden by settings in stephandle}
}
\description{
finishRun
}
\references{
ics1140
}
