% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleFromStep.R
\name{fullProcess}
\alias{fullProcess}
\title{fullProcess
collect all data for one process
cycle with old version. test usage of correct run, with revision}
\usage{
fullProcess(process, stepHandle)
}
\arguments{
\item{process}{the process loaded from the repo}

\item{stepHandle}{the stepHandle the process will be added to}
}
\description{
fullProcess
collect all data for one process
cycle with old version. test usage of correct run, with revision
}
