% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportImportUtils.R
\name{generateLinkMappingTemplate}
\alias{generateLinkMappingTemplate}
\title{Generate link mapping template from outside links}
\usage{
generateLinkMappingTemplate(allOutsideLinks)
}
\arguments{
\item{allOutsideLinks}{Data frame of all outside links (full version, with ident/version/filehash/name).}
}
\value{
Data frame of link mappings, or NULL if no links.
}
\description{
Generate link mapping template from outside links
}
\keyword{internal}
