% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportImportUtils.R
\name{generateToolMappingTemplate}
\alias{generateToolMappingTemplate}
\title{Generate tool mapping template from workflow steps}
\usage{
generateToolMappingTemplate(workflowDf)
}
\arguments{
\item{workflowDf}{Data frame of workflow steps.}
}
\value{
Data frame of tool mappings with key, runserverLabel, toolLabel, toolInstance, gridTool.
}
\description{
Generate tool mapping template from workflow steps
}
\keyword{internal}
