% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{getAbstract}
\alias{getAbstract}
\title{abstract function for loading}
\usage{
getAbstract(
  ident,
  from = pwd(),
  addAsLink = TRUE,
  addIdToName = TRUE,
  caption = "",
  description = "",
  folderName,
  func,
  ...
)
}
\arguments{
\item{ident}{path, resource or entity ID of the picture}

\item{from}{used for relative paths, by default pwd is used, which is initiated with the step that started improveR}

\item{addAsLink}{creates a link in inventory if TRUE, warn improveClean has to be run at the end}

\item{addIdToName}{adds the entity ID to the filename, enity version ID if a specific version is retrieved}

\item{caption}{by default entityID and lastmodified are the caption, here alternative text can be provided}

\item{description}{by default filename is the caption, here alternative text can be provided}

\item{folderName}{where to store the file}

\item{func}{the processing function for this type}

\item{...}{pass over arguments for the read function}
}
\description{
abstract function for loading
}
