% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenRefresherPlugin.R
\name{getActiveTokenRefresher}
\alias{getActiveTokenRefresher}
\title{Get active token refresher}
\usage{
getActiveTokenRefresher()
}
\value{
Token refresher object or NULL
}
\description{
Gets the currently active token refresher
}
