% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{getCopy}
\alias{getCopy}
\title{gets a file objectas local copy}
\usage{
getCopy(ident, from = pwd(), caption = "", description = "")
}
\arguments{
\item{ident}{path, resource or entity ID of the file}

\item{from}{used for relative paths, by default pwd is used, which is initiated with the step that started improveR}

\item{caption}{by default entityID and lastmodified are the caption, here alternative text can be provided}

\item{description}{by default filename is the caption, here alternative text can be provided}
}
\description{
gets a file objectas local copy
}
\references{
ics1141
}
