% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrectId.R
\name{getCorrectId}
\alias{getCorrectId}
\title{getCorrectId}
\usage{
getCorrectId(resolveToId)
}
\arguments{
\item{resolveToId}{the value resolve to short entity ID, or resource ID, version possible for both}
}
\description{
Resolves various ident types to an ID.
\itemize{
\item a resource => resource ID
\item a resource ID => resource ID
\item a resource versionID => resource version ID
\item a long entity ID with http: ... => short entity ID
\item a short entity ID => short entity ID
\item an entity ID without prefix => short entity ID
}
}
\examples{
\dontrun{
getCorrectId("112EE78F4CDC4400836F8C059AF2EA5F") #resourceId
getCorrectId("02C347E7439942FE834C7714F49EF082") #resourceVersionId
getCorrectId("http://host/provide?resourceId=resourceId=my_server:ST-63657") #long entity id
getCorrectId("my_server:ST-63657") #short entity id
}
}
\references{
ics1087
}
