% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenManagement.R
\name{getCurrentTokenData}
\alias{getCurrentTokenData}
\title{Get current token data}
\usage{
getCurrentTokenData()
}
\value{
List with token data or NULL if no token
}
\description{
Gets the current token data from environment variables
}
