% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getDataByResource}
\alias{getDataByResource}
\title{retrieves data table for rMarkdown, .xls and .xlsx files assumed to be excel, .rds as RDS, .sas7bdat as SAS all others as CSV}
\usage{
getDataByResource(
  resourceDesc,
  caption = "",
  description = "",
  parser = NULL,
  ...
)
}
\arguments{
\item{resourceDesc}{resource descriptor of the data file}

\item{caption}{by default entityID and lastmodified are the caption, here alternative text can be provided}

\item{description}{by default filename is the caption, here alternative text can be provided}

\item{parser}{a custom parsing function can be handed over. It needs to take a local path as first argument}

\item{...}{pass over arguments for the read function}
}
\description{
retrieves data table for rMarkdown, .xls and .xlsx files assumed to be excel, .rds as RDS, .sas7bdat as SAS all others as CSV
}
