% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{getDesc}
\alias{getDesc}
\title{builds a default caption and description}
\usage{
getDesc(resourceDesc, caption = "", description = "")
}
\arguments{
\item{resourceDesc}{resource descriptor of the picture}

\item{caption}{by default entityID and lastmodified are the caption, here alternative text can be provided}

\item{description}{by default filename is the caption, here alternative text can be provided}
}
\description{
builds a default caption and description
}
