% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTreeConnections.R
\name{getFullUsage}
\alias{getFullUsage}
\title{getFullUsage
returns a workflowHandle for the usage workflow of a step, covering multiple trees.}
\usage{
getFullUsage(ident, from = pwd(), workflowEnv = NULL, depth = -1)
}
\arguments{
\item{ident}{ident of the step}

\item{from}{from for relative pathes, default is pwd()}

\item{workflowEnv}{internal variable, leave NULL}

\item{depth, }{the depth of recursion when retrieving additional trees. -1 takes all trees into account}
}
\description{
getFullUsage
returns a workflowHandle for the usage workflow of a step, covering multiple trees.
}
