% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{getHandleForResource}
\alias{getHandleForResource}
\title{getHandleForResource
returns the stepHandle for a specific step resource.
Only works if a prepared step has been executed or a workflow was pulled from the repository and not detached via detachWorkflowFromResources}
\usage{
getHandleForResource(workflowHandle, ident, from = pwd())
}
\arguments{
\item{workflowHandle}{the workflow to seach for the step}

\item{ident}{ident of the resource}

\item{from, }{path for relative pathes, default pwd()}
}
\description{
getHandleForResource
returns the stepHandle for a specific step resource.
Only works if a prepared step has been executed or a workflow was pulled from the repository and not detached via detachWorkflowFromResources
}
