% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{getLatestRevision}
\alias{getLatestRevision}
\title{Get Latest Revision}
\usage{
getLatestRevision()
}
\value{
A list with the latest revision data, or \code{NULL} if not available.
}
\description{
Retrieves the latest revision (transaction) from the repository.
}
\references{
ccs10
}
