% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDetails.R
\name{getLatestRun}
\alias{getLatestRun}
\title{Get Latest Run}
\usage{
getLatestRun(ident, processId)
}
\arguments{
\item{ident}{Step identifier (path, resource ID, entity ID, etc.).}

\item{processId}{The ID of the process.}
}
\value{
A data frame with the latest run data, or \code{NULL} if not found.
}
\description{
Retrieves the latest run for a process belonging to a step.
}
\references{
ics1329
}
