% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLineage.R
\name{getLineage_deprecated}
\alias{getLineage_deprecated}
\title{Get lineage for version < 4.4}
\usage{
getLineage_deprecated(step, env)
}
\description{
Gets all links in the step, finds their targets, then traverses up
the resource hierarchy to find which steps contain those targets.
}
\keyword{internal}
