% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTreeConnections.R
\name{getLineageLinkedStepsinOtherTrees}
\alias{getLineageLinkedStepsinOtherTrees}
\title{getLineageLinkedStepsinOtherTrees
returns all steps in other trees that are referenced via link from a specific step}
\usage{
getLineageLinkedStepsinOtherTrees(ident, from = pwd())
}
\arguments{
\item{ident}{ident of the step}

\item{from}{from for relative pathes, default is pwd()}
}
\description{
getLineageLinkedStepsinOtherTrees
returns all steps in other trees that are referenced via link from a specific step
}
