% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliSetup.R
\name{getLocalRepoResource}
\alias{getLocalRepoResource}
\title{Get Local Repository Resource}
\usage{
getLocalRepoResource(localPath)
}
\arguments{
\item{localPath}{The local repository path.}
}
\value{
The resource information.
}
\description{
Retrieves2 the resource information for a local repository.
}
