% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParent.R
\name{getParent}
\alias{getParent}
\title{returns the parent of anything resolved by getCorrectId}
\usage{
getParent(identifier)
}
\arguments{
\item{identifier, }{ID or resource}
}
\description{
returns the parent of anything resolved by getCorrectId
}
\references{
ics1088
}
