% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{getReviewEntryComments}
\alias{getReviewEntryComments}
\title{Get Review Etnry Comments}
\usage{
getReviewEntryComments(resourceId, entryId, from = pwd())
}
\arguments{
\item{resourceId}{id (UUID) of the resource}

\item{entryId}{id (UUID) of the review entry}

\item{from}{defaults to pwd, used to resolve relative paths}
}
\description{
Get Review Etnry Comments
}
\references{
ics1543
}
