% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{getReviewers}
\alias{getReviewers}
\title{Get Reviewers}
\usage{
getReviewers(ident, from = pwd())
}
\arguments{
\item{ident}{id}

\item{from}{defaults to pwd, used to resolve relative paths}
}
\description{
Get Reviewers
}
\references{
ics1208
}
