% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDetails.R
\name{getRun}
\alias{getRun}
\title{Get Run}
\usage{
getRun(ident, processId, runId)
}
\arguments{
\item{ident}{Step identifier (path, resource ID, entity ID, etc.).}

\item{processId}{The ID of the process.}

\item{runId}{The ID of the run.}
}
\value{
A data frame with the run data, or \code{NULL} if not found.
}
\description{
Retrieves a specific run by its ID for a process belonging to a step.
}
\references{
ics1330
}
