% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDetails.R
\name{getRunGridArguments}
\alias{getRunGridArguments}
\title{Get Run Grid Arguments}
\usage{
getRunGridArguments(ident, processId, runId)
}
\arguments{
\item{ident}{Step identifier (path, resource ID, entity ID, etc.).}

\item{processId}{The ID of the process.}

\item{runId}{The ID of the run.}
}
\value{
A data frame with the run grid arguments, or \code{NULL} if not found.
}
\description{
Retrieves the grid arguments of a specific run for a process belonging to a step.
}
\references{
ics1332
}
