% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDetails.R
\name{getRunPhases}
\alias{getRunPhases}
\title{Get Run Phases}
\usage{
getRunPhases(ident, processId, runId)
}
\arguments{
\item{ident}{Step identifier (path, resource ID, entity ID, etc.).}

\item{processId}{The ID of the process.}

\item{runId}{The ID of the run.}
}
\value{
A data frame with the run phases, or \code{NULL} if not found.
}
\description{
Retrieves the phases of a specific run for a process belonging to a step.
}
\references{
ics1333
}
