% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedTokenManager.R
\name{getSharedTokenFilePath}
\alias{getSharedTokenFilePath}
\title{Get shared token file path}
\usage{
getSharedTokenFilePath()
}
\value{
Character string with path to shared token file
}
\description{
Gets the fixed path for shared encrypted tokens
}
