% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStep.R
\name{getStepResource}
\alias{getStepResource}
\title{getStepResource}
\usage{
getStepResource(handle)
}
\arguments{
\item{handle}{load the step resource if the handle has already been executed}
}
\description{
getStepResource
}
\references{
ics1221
}
