% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_realise.R
\name{getStepState_deprecated}
\alias{getStepState_deprecated}
\title{Get step state for deprecated version}
\usage{
getStepState_deprecated(step)
}
\description{
Get step state for deprecated version
}
\keyword{internal}
