% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getTextByResource}
\alias{getTextByResource}
\title{retrieves text string for rMarkdown}
\usage{
getTextByResource(resourceDesc, caption = "", description = "", ...)
}
\arguments{
\item{resourceDesc}{resource descriptor of the picture}

\item{caption}{by default entityID and lastmodified are the caption, here alternative text can be provided}

\item{description}{by default filename is the caption, here alternative text can be provided}

\item{...}{pass over arguments for the read function}
}
\description{
retrieves text string for rMarkdown
}
