% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getTextString}
\alias{getTextString}
\title{gets a text file as chracter string and retrieves the caption and description by the resource ID or entity ID}
\usage{
getTextString(
  ident,
  from = pwd(),
  addAsLink = TRUE,
  caption = "",
  description = "",
  ...
)
}
\arguments{
\item{ident}{path, resource or entity ID of the picture}

\item{from}{used for relative paths, by default pwd is used, which is initiated with the step that started improveR}

\item{addAsLink}{creates a link in inventory if TRUE, warn improveClean has to be run at the end}

\item{caption}{by default entityID and lastmodified are the caption, here alternative text can be provided}

\item{description}{by default filename is the caption, here alternative text can be provided}

\item{...}{pass over arguments for the read function}
}
\description{
gets a text file as chracter string and retrieves the caption and description by the resource ID or entity ID
}
\references{
ics1141
}
