% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsage.R
\name{getUsage_deprecated}
\alias{getUsage_deprecated}
\title{Get usage for version < 4.4}
\usage{
getUsage_deprecated(step, env)
}
\description{
Gets all files in the step, loads their references, then traverses up
the resource hierarchy to find which steps contain those references.
}
\keyword{internal}
