% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTreeConnections.R
\name{getUsageLinkedStepsinOtherTrees}
\alias{getUsageLinkedStepsinOtherTrees}
\title{getUsageLinkedStepsinOtherTrees
returns all steps in other trees that reference a specific step via link}
\usage{
getUsageLinkedStepsinOtherTrees(ident, from = pwd())
}
\arguments{
\item{ident}{ident of the step}

\item{from}{from for relative pathes, default is pwd()}
}
\description{
getUsageLinkedStepsinOtherTrees
returns all steps in other trees that reference a specific step via link
}
