% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWorkflow.R
\name{getWorkflow}
\alias{getWorkflow}
\title{getWorkflow
reads a tree and creates stepEnvs for all steps}
\usage{
getWorkflow(ident, from = pwd(), includeSelf = F)
}
\arguments{
\item{ident}{the ident of the step}

\item{from}{if relative path, default the starting step}

\item{includeSelf}{by defaults does not include the step the script is run with.}
}
\description{
getWorkflow
reads a tree and creates stepEnvs for all steps
}
