% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlesFromTree.R
\name{handlesFromTree}
\alias{handlesFromTree}
\title{handlesFromTree
reads a tree and creates handles for all steps}
\usage{
handlesFromTree(ident, from = pwd(), includeSelf = F)
}
\arguments{
\item{ident}{the ident of the step}

\item{from}{if relative path, default the starting step}

\item{includeSelf}{by defaults does not include the step the script is run with.}
}
\description{
handlesFromTree
reads a tree and creates handles for all steps
}
