% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.R
\name{hasCurrentError}
\alias{hasCurrentError}
\title{hasCurrentError checks if there is an error message in the current context}
\usage{
hasCurrentError()
}
\description{
hasCurrentError checks if there is an error message in the current context
}
