% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sshTunnel.R
\name{improveCloseTunnel}
\alias{improveCloseTunnel}
\title{Close an SSH Tunnel}
\usage{
improveCloseTunnel(tunnel)
}
\arguments{
\item{tunnel}{A tunnel object returned by \code{\link{improveOpenTunnel}}.}
}
\value{
\code{NULL}, invisibly. Stops with an error if \code{tunnel} is not the object
\code{\link[=improveOpenTunnel]{improveOpenTunnel()}} returned.
}
\description{
Closes an SSH tunnel previously opened by \code{\link{improveOpenTunnel}}.
}
\examples{
\dontrun{
tunnel <- improveOpenTunnel(step)
# ... use RStudio ...
improveCloseTunnel(tunnel)
}

}
\seealso{
\code{\link{improveOpenTunnel}}
}
