% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveConnect.R
\name{improveConnected}
\alias{improveConnected}
\title{improveConnected}
\usage{
improveConnected(silent = FALSE)
}
\arguments{
\item{silent}{Should logging messages be hidden. Default is FALSE.}
}
\value{
logical. Returns \code{TRUE} if the connection is established, \code{FALSE} otherwise. The value is returned invisibly.
}
\description{
improveConnected checks if improveConnect was called.
}
\seealso{
\code{\link[=improveConnect]{improveConnect()}}, \code{\link[=improveDisconnect]{improveDisconnect()}}
}
