% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveConnect.R
\name{improveDisconnect}
\alias{improveDisconnect}
\title{improveDisconnect}
\usage{
improveDisconnect(env = cacheEnv)
}
\arguments{
\item{env}{default is 'cacheEnv'}
}
\description{
improveDisconnect removes all connection information.
}
\seealso{
\code{\link[=improveConnect]{improveConnect()}}, \code{\link[=improveConnected]{improveConnected()}}
}
