% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editable.R
\name{improveEditable}
\alias{improveEditable}
\title{improveEditable}
\usage{
improveEditable()
}
\description{
improveEditable checks if improveConnect was called and if the editable flag was set
}
\seealso{
\code{\link[=improveConnected]{improveConnected()}}
}
