% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{improveLogin}
\alias{improveLogin}
\title{improveLogin}
\usage{
improveLogin(
  repo,
  user,
  password,
  shortEntityId,
  logLevel = "INFO",
  secure = T
)
}
\arguments{
\item{repo}{the repository URL in this form https://\if{html}{\out{<url>}}:\if{html}{\out{<port>}}/\if{html}{\out{<repositoryPath>}} the api part (/api/v1) is added automatically}

\item{user}{the username as character}

\item{password}{the plain text password as character, is not logged or stored}

\item{shortEntityId}{one valid short entity ID must be provided, this is used as pathWorkingDirectory}

\item{logLevel}{possible LogLevels: DEBUG, INFO, WARN, ERROR}

\item{secure}{if TRUE the certificates are checked}
}
\description{
improveLogin can be used for test, development, or interactive uses of improveR.
It uses the username and password to retrieve a token.
}
\references{
ics1081
}
