% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauthLogin.R
\name{improveOAuth}
\alias{improveOAuth}
\title{improveOAuth}
\usage{
improveOAuth(
  repo,
  shortEntityId = "/",
  logLevel = "INFO",
  secure = T,
  openBrowser = T,
  withCodeVerifier = T
)
}
\arguments{
\item{repo}{the repository URL in this form https://\if{html}{\out{<url>}}:\if{html}{\out{<port>}}/\if{html}{\out{<repositoryPath>}} the api part (/api/v1) is added automatically}

\item{shortEntityId}{one valid short entity ID must be provided, this is used as pathWorkingDirectory}

\item{logLevel}{possible LogLevels: DEBUG, INFO, WARN, ERROR}

\item{secure}{if TRUE the certificates are checked.
Default is TRUE, it can be set to false also by the environment variable IMPROVER_SECURITY=insecure}

\item{openBrowser}{this indicates wether the r session can open a browser to user can access. default is true. if set to false a URL and a code is written to the console. This can be used to log in form a different PC.
it can be set to false also by the environment variable IMPROVER_HEADLESS_OAUTH=T}

\item{withCodeVerifier}{if the oauth provider uses pkca code challenge verification}
}
\description{
improveOAuth is used to connect to the repository via oauth.
}
\references{
ics1081
}
