% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveR-package.R
\docType{package}
\name{improveR-package}
\alias{improveR}
\alias{improveR-package}
\title{improveR: Readonly package for improveR}
\description{
unifies all readonly improveR functions.
}
\author{
\strong{Maintainer}: Michael Hackl \email{michael.hackl@scinteco.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
