% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{improveReLogin}
\alias{improveReLogin}
\title{improveReLogin}
\usage{
improveReLogin(password)
}
\arguments{
\item{password}{The plain password as a character string. The password is not logged or stored.}
}
\description{
improveReLogin can be used if a token could not be refreshed for test, development, or interactive uses
of improveR. The function uses the password to retrieve a token. The user has to have logged in before with improveLogin.
}
\references{
ics1208
}
