% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliNew.R
\name{infoCli}
\alias{infoCli}
\title{Repository Info}
\usage{
infoCli(localPath)
}
\arguments{
\item{localPath}{The local repository path.}
}
\value{
CLI output with repository info.
}
\description{
Displays information about a local repository and its server-side root
resource (profile, server URL, user, path, IDs).
}
