% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenRefresherPlugin.R
\name{initializeTokenRefreshers}
\alias{initializeTokenRefreshers}
\title{Initialize default token refreshers}
\usage{
initializeTokenRefreshers()
}
\description{
Registers the default token refresher implementations
}
