% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRun.R
\name{initRepo}
\alias{initRepo}
\title{set runName and testServer to environment}
\usage{
initRepo(runName = "test1", testServer = "RobertEnv")
}
\arguments{
\item{runName}{name for the result folder}

\item{testServer}{a registered server}
}
\description{
arguments:
}
