% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedTokenManager.R
\name{isSharedTokenRefreshRunning}
\alias{isSharedTokenRefreshRunning}
\title{Check if shared token refresh is running}
\usage{
isSharedTokenRefreshRunning()
}
\value{
Logical indicating if shared tokens are available
}
\description{
Checks if the shared encrypted token file exists
}
