% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerCloseFunction.R
\name{listCloseFunctions}
\alias{listCloseFunctions}
\title{listCloseFunctions}
\usage{
listCloseFunctions()
}
\description{
Lists the names of all functions registered by registerCloseFunction.
}
\seealso{
\code{\link[=registerCloseFunction]{registerCloseFunction()}}
}
