% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadChildResources.R
\name{loadChildResources}
\alias{loadChildResources}
\title{loadChildResources}
\usage{
loadChildResources(ident, from = pwd())
}
\arguments{
\item{ident}{the resource id or the entity id of the resource; can be a list}

\item{from}{used if a relative path is used}
}
\description{
Loads all child resources by the resourceId, entity ID, or entity version ID.
Uses caching. Child resources can be folders, steps, workflows, or analysis trees.
The results are returned as a data frame or a list of data frames.
The dates are  converted to POSIX dates with the convertImproveTimestampToPosix function.
}
\seealso{
\code{\link[=convertImproveTimestampToPosix]{convertImproveTimestampToPosix()}}
}
