% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGridArguments.R
\name{loadGridArgumentDefinition}
\alias{loadGridArgumentDefinition}
\title{loadGridArgumentDefinition loads grid argument by gridProvider and name}
\usage{
loadGridArgumentDefinition(gridProvider, argumentName)
}
\arguments{
\item{gridProvider}{label of the grid provider}

\item{argumentName}{name of the argument}
}
\description{
loadGridArgumentDefinition loads grid argument by gridProvider and name
}
\references{
ics1216
}
