% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGridArguments.R
\name{loadGridArguments}
\alias{loadGridArguments}
\title{loadGridArguments}
\usage{
loadGridArguments(gridProvider)
}
\arguments{
\item{gridProvider}{label of the grid provider}
}
\description{
loadGridArguments
}
\references{
ics1216
}
