% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetaDataDefinitions.R
\name{loadMetaDataDefinitions}
\alias{loadMetaDataDefinitions}
\title{Loads All Meta Data Definitions For A Scope, The Default Scope Is "Improve Client"}
\usage{
loadMetaDataDefinitions(scope = "Improve Client")
}
\arguments{
\item{scope}{the metadata scope}
}
\description{
Loads All Meta Data Definitions For A Scope, The Default Scope Is "Improve Client"
}
\references{
ics1137
}
