% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProcess.R
\name{loadProcessesForStep}
\alias{loadProcessesForStep}
\title{loadProcessesForStep}
\usage{
loadProcessesForStep(stepIdent)
}
\arguments{
\item{stepIdent}{resourceId of the step or the step}
}
\description{
loadProcessesForStep
}
\references{
ics1218
}
