% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProcess.R
\name{loadProcessesForStepById}
\alias{loadProcessesForStepById}
\title{loadProcessesForStepById}
\usage{
loadProcessesForStepById(processId)
}
\arguments{
\item{processId}{processId of the process}
}
\description{
loadProcessesForStepById
}
\references{
ics1218
}
