% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProcess.R
\name{loadProcessGridArguments}
\alias{loadProcessGridArguments}
\title{loadProcessGridArguments loads the tool for a given process}
\usage{
loadProcessGridArguments(processId)
}
\arguments{
\item{processId}{id of the process}
}
\description{
loadProcessGridArguments loads the tool for a given process
}
\references{
ics1218
}
