% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProcess.R
\name{loadProcessRuns}
\alias{loadProcessRuns}
\title{loadProcessRun loads the runs for a given process}
\usage{
loadProcessRuns(processId)
}
\arguments{
\item{processId}{id of the process}
}
\description{
loadProcessRun loads the runs for a given process
}
\references{
ics1218
}
