% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProcess.R
\name{loadProcessVariables}
\alias{loadProcessVariables}
\title{loadProcessVariables loads the variables for a given process}
\usage{
loadProcessVariables(processId)
}
\arguments{
\item{processId}{id of the process}
}
\description{
loadProcessVariables loads the variables for a given process
}
\references{
ics1218
}
