% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceRelations.R
\name{loadRelationTypes}
\alias{loadRelationTypes}
\title{loads all registered relation types}
\usage{
loadRelationTypes(conn)
}
\arguments{
\item{conn}{database connection}
}
\description{
loads all registered relation types
}
