% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceByPathGeneric.R
\name{loadResourceByPathGeneric}
\alias{loadResourceByPathGeneric}
\title{loadResourceByPathGeneric}
\usage{
loadResourceByPathGeneric(path, from = pwd())
}
\arguments{
\item{path}{path to the resource.}

\item{from}{working directory from which relative path is resolved.
Default is the root directory.}
}
\description{
Loads a resource by its path.
}
