% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepConfigData.R
\name{loadToolForRunserver}
\alias{loadToolForRunserver}
\title{loadToolForRunserver}
\usage{
loadToolForRunserver(runserverId, toolName = NULL, toolInstanceName = NULL)
}
\arguments{
\item{runserverId}{resourceId of the runserver}

\item{toolName}{name of the tool, optional, but toolname or toolInstanceName have to be given}

\item{toolInstanceName}{name of the tool instance, optional, but toolname or toolInstanceName have to be given}
}
\description{
loadToolForRunserver
}
\references{
ics1227
}
