% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepConfigData.R
\name{loadToolsForRunserver}
\alias{loadToolsForRunserver}
\title{loadToolsForRunserver}
\usage{
loadToolsForRunserver(runserverId)
}
\arguments{
\item{runserverId}{resourceId of the runserver}
}
\description{
loadToolsForRunserver
}
\references{
ics1227
}
