% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexecute.R
\name{makeStepsRelative}
\alias{makeStepsRelative}
\title{makeStepsRelative
takes all steps in a list of steps and makes them relative to each other}
\usage{
makeStepsRelative(workflow)
}
\arguments{
\item{workflow}{a dataframe of stepHandles that have already been executed}
}
\description{
makeStepsRelative
takes all steps in a list of steps and makes them relative to each other
}
\references{
ics1220
}
