% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{move}
\alias{move}
\title{moves resources
multiple sources can be moved at once
if multiple sources are moved target needs to be a container and targetName empty
if only one source is moved, a new name can be specified, this can be done via a path or via targetName}
\usage{
move(
  sources,
  target,
  targetName = "",
  overwrite = F,
  comment = "modified by improveRW"
)
}
\arguments{
\item{sources}{the files to be moved}

\item{target}{the target folder}

\item{targetName}{name for the moved file}

\item{overwrite}{flag if a file already existing at the target location should be overwritten if existing. If flag is false and file exists NULL is returned}

\item{comment}{comment for the commit, defaults to "modified by improveRW"}
}
\description{
moves resources
multiple sources can be moved at once
if multiple sources are moved target needs to be a container and targetName empty
if only one source is moved, a new name can be specified, this can be done via a path or via targetName
}
\references{
ics1139
}
