% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportImportUtils.R
\name{moveFile}
\alias{moveFile}
\title{Move files between directories with cross-device fallback}
\usage{
moveFile(sourcePath, destPath)
}
\arguments{
\item{sourcePath}{Source file path.}

\item{destPath}{Destination file path.}
}
\description{
Move files between directories with cross-device fallback
}
\keyword{internal}
