% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisePath.R
\name{normalisePath}
\alias{normalisePath}
\title{normalisePath}
\usage{
normalisePath(path, startPath = "/")
}
\arguments{
\item{path}{the path to normalise}

\item{startPath}{Starting point for relative pathes, defaults to /}
}
\description{
normalisePath converts any path to an absolute path, and eleminates all .. and .
If a relative path is given, it needs an absolute path as startPath to resolve.
If by too many .. the path navigates beyond the root, NULL is returned.
}
\examples{
\dontrun{
improveR::normalisePath(path = "./../lmer/../lmer", startPath = "/0demo/lmer") # /0demo/lmer
improveR::normalisePath(path = "./../../../lmer/../lmer", startPath = "/0demo/lmer") # NULL
improveR::normalisePath(path = "/0demo/lmer", startPath = "/0demo/lmer") # /0demo/lmer
}
}
\references{
ics1089
}
