% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worflowOrder.R
\name{orderedLineage}
\alias{orderedLineage}
\title{retrieves a data frame holding the lineage of a step within the given workflow.
One workflow can span multiple analsis trees
the order of the steps is a possible execution order. If no execution order can be built, NULL is returned}
\usage{
orderedLineage(workflow, stepHandle)
}
\arguments{
\item{workflow}{as df}

\item{stepHandle}{id of the lineage step}
}
\description{
retrieves a data frame holding the lineage of a step within the given workflow.
One workflow can span multiple analsis trees
the order of the steps is a possible execution order. If no execution order can be built, NULL is returned
}
\references{
ics1231
}
