% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worflowOrder.R
\name{orderedUsage}
\alias{orderedUsage}
\title{retrieves a data frame holding the usage of a step within the given workflow.
One workflow can span multiple analsis trees
the order of the steps is a possible execution order}
\usage{
orderedUsage(workflow, stepHandle)
}
\arguments{
\item{workflow}{as df}

\item{stepHandle}{id of the lineage step}
}
\description{
retrieves a data frame holding the usage of a step within the given workflow.
One workflow can span multiple analsis trees
the order of the steps is a possible execution order
}
\references{
ics1231
}
