% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlesFromTree.R
\name{persistWorkflowChanges}
\alias{persistWorkflowChanges}
\title{persistWorkflowChanges
takes all changes made to the data frame and persists them with their step handle}
\usage{
persistWorkflowChanges(workflowDf)
}
\arguments{
\item{workflowDf}{the workflow dataframe as retrieved by retrieveWorkflow}
}
\description{
persistWorkflowChanges
takes all changes made to the data frame and persists them with their step handle
}
\references{
ics1215
}
