% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.R
\name{printLogs}
\alias{printLogs}
\title{printLogs}
\usage{
printLogs(reset = FALSE)
}
\arguments{
\item{reset}{Boolean, TRUE resets the list after printing.}
}
\description{
printLogs prints the strucured log list to the console.
}
