% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProcess.R
\name{processGridProvider}
\alias{processGridProvider}
\title{processGridProvider loads the grid provider for a given process}
\usage{
processGridProvider(process)
}
\arguments{
\item{process}{the data frame for the process}
}
\description{
processGridProvider loads the grid provider for a given process
}
\references{
ics1218
}
