% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProcess.R
\name{processTool}
\alias{processTool}
\title{processTool loads the tool for a given process}
\usage{
processTool(process)
}
\arguments{
\item{process}{the data frame for the process}
}
\description{
processTool loads the tool for a given process
}
\references{
ics1218
}
