% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{queryFolder}
\alias{queryFolder}
\title{performs this query against the elastic search within one folder}
\usage{
queryFolder(
  queryString,
  fetchSize = 100,
  fetchOffset = 0,
  ident,
  fromForRelativePathes = improveRcore::pwd()
)
}
\arguments{
\item{queryString}{query with the same grammar as in the UI}

\item{fetchSize}{maximum number of returned results}

\item{fetchOffset}{offset of the results, enables paging}

\item{ident}{path or id of a resource to start search from}

\item{fromForRelativePathes}{root for a relative path per default pwd}
}
\description{
performs this query against the elastic search within one folder
}
