% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliSetup.R
\name{readLocalRepoInfo}
\alias{readLocalRepoInfo}
\title{Read Local Repository Info}
\usage{
readLocalRepoInfo(localPath)
}
\arguments{
\item{localPath}{The local repository path.}
}
\value{
A list of repository properties.
}
\description{
Reads the repository properties file for a local repository.
}
