% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedTokenManager.R
\name{readSharedRefreshedTokens}
\alias{readSharedRefreshedTokens}
\title{Read shared refreshed tokens}
\usage{
readSharedRefreshedTokens(secret = NULL)
}
\arguments{
\item{secret}{Optional. The shared secret. If NULL, gets from environment}
}
\value{
Logical indicating success
}
\description{
Reads and applies tokens from shared encrypted file
}
