% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.R
\name{redirectLogs}
\alias{redirectLogs}
\title{redirectLogs}
\usage{
redirectLogs(redirect)
}
\arguments{
\item{redirect}{Boolean, TRUE turns on redirection,  FALSE turns it off.}
}
\description{
redirectLogs redirects the log from the default logstream to a structured list, with a different context.
}
