% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetaDataDefinitions.R
\name{refreshMetaDataDefinitions}
\alias{refreshMetaDataDefinitions}
\alias{updateMetaDataDefinitions}
\title{Update Meta Data Definitions}
\usage{
refreshMetaDataDefinitions(...)


updateMetaDataDefinitions(scope = "Improve Client")
}
\arguments{
\item{scope}{the metadata scope}
}
\description{
\code{refreshMetaDataDefinitions} is deprecated; use \code{refreshMetaDataDefinitions} instead.

Update Meta Data Definitions
}
\references{
ics1137
}
