% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceRelations.R
\name{refreshResourceRelations}
\alias{refreshResourceRelations}
\alias{updateResourceRelations}
\title{Reloads the Resource Relations}
\usage{
refreshResourceRelations(...)


updateResourceRelations(ident, from = pwd())
}
\arguments{
\item{ident}{Identifier of the resource.}

\item{from}{Base path for resolving relative paths. Defaults to \code{pwd()}.}
}
\description{
\code{refreshResourceRelations} is deprecated; use \code{refreshResourceRelations} instead.

Reloads the Resource Relations
}
\references{
ics1044
}
