% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerCloseFunction.R
\name{registerCloseFunction}
\alias{registerCloseFunction}
\title{registerCloseFunction registers a function that is called when the r session is ended. In a batch job the r session is ended when the run is over, in rstudio the session is ended, when the user either closes rstudio with the close button or selects terminate or restart session
the close functions are executed in alphabetical order, so numbering as prefix can guarantee an order}
\usage{
registerCloseFunction(name, func)
}
\arguments{
\item{name}{name of the function, this name is used for the alphabetical ordering and unregistering}

\item{func, }{the function that is executed}
}
\description{
registerCloseFunction registers a function that is called when the r session is ended. In a batch job the r session is ended when the run is over, in rstudio the session is ended, when the user either closes rstudio with the close button or selects terminate or restart session
the close functions are executed in alphabetical order, so numbering as prefix can guarantee an order
}
