% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favorites.R
\name{removeFavorite}
\alias{removeFavorite}
\title{Remove Favorite}
\usage{
removeFavorite(resourceId)
}
\arguments{
\item{resourceId}{Resource ID of the favorite to remove.}
}
\value{
\code{TRUE} if the favorite was removed successfully, \code{FALSE} otherwise.
}
\description{
Removes a resource from the user's favorites collection.
}
\references{
ics1800
}
