% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{removeStepLocalFile}
\alias{removeStepLocalFile}
\title{removes a local file from the step}
\usage{
removeStepLocalFile(stepHandle, name = NULL, path = NULL)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{name}{the name of the file in the inventory}

\item{path}{the path of the file in the filesystem}
}
\description{
removes a local file from the step
}
