% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{removeStepRemoteFile}
\alias{removeStepRemoteFile}
\title{removes a remote file from the step}
\usage{
removeStepRemoteFile(stepHandle, name = NULL, ident = NULL)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{name}{the name of the file in the inventory}

\item{ident}{the ident of the file in the inventory}
}
\description{
removes a remote file from the step
}
