% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexecute.R
\name{rerunChangedAndOutdated}
\alias{rerunChangedAndOutdated}
\title{rerunChangedAndOutdated}
\usage{
rerunChangedAndOutdated(
  ident = pwd()$parentId,
  from = pwd(),
  includeUsage = T,
  force = F,
  lastSteps = NULL
)
}
\arguments{
\item{ident}{the tree that needs to be rerun}

\item{from}{pwd for defining the tree relativ}

\item{includeUsage}{include the usage of the steps}

\item{force}{execute if no files changed and no link outdated}

\item{lastSteps}{a list of entityIDs of steps that should be executed in the end. Those steps may not have usage}
}
\description{
rerunChangedAndOutdated
}
\references{
ics1219
}
