% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexecute.R
\name{rerunSteps}
\alias{rerunSteps}
\title{rerunSteps}
\usage{
rerunSteps(
  steps,
  from = pwd()$parentId,
  includeLineage = F,
  includeUsage = T,
  force = F
)
}
\arguments{
\item{steps}{list of steps that have to be rerun}

\item{from}{the tree the steps are in}

\item{includeLineage}{include the lineage within the tree}

\item{includeUsage}{include the usage of the steps}

\item{force}{execute if no files changed and no link outdated}
}
\description{
rerunSteps
}
\references{
ics1219
}
