% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexecute.R
\name{rerunTrees}
\alias{rerunTrees}
\title{rerunTrees reruns one or more trees, reexecutes all steps}
\usage{
rerunTrees(trees, from = pwd()$parentId)
}
\arguments{
\item{trees}{list of steps that have to be rerun}

\item{from}{the tree the steps are in}
}
\description{
rerunTrees reruns one or more trees, reexecutes all steps
}
\references{
ics1219
}
