% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{resetStep}
\alias{resetStep}
\title{resets the step by deleting all the downloaded data, if you are using relative pathes it makes sense to do this at the beginning of the step
it also deletes all the links that were currently created}
\usage{
resetStep()
}
\description{
resets the step by deleting all the downloaded data, if you are using relative pathes it makes sense to do this at the beginning of the step
it also deletes all the links that were currently created
}
