% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportImportUtils.R
\name{resolveInternalLinksSecondPass}
\alias{resolveInternalLinksSecondPass}
\title{Resolve internal links in second pass after all steps are created}
\usage{
resolveInternalLinksSecondPass(internalLinks, importWF)
}
\arguments{
\item{internalLinks}{Data frame of internal links.}

\item{importWF}{Data frame with fullName and newEntityId columns.}
}
\description{
Resolve internal links in second pass after all steps are created
}
\keyword{internal}
