% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportImportUtils.R
\name{restoreParentRelationships}
\alias{restoreParentRelationships}
\title{Restore parent relationships after all steps are created}
\usage{
restoreParentRelationships(importWF)
}
\arguments{
\item{importWF}{Data frame with parentFullName, fullName, and newEntityId columns.}
}
\description{
Restore parent relationships after all steps are created
}
\keyword{internal}
