% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStepEnv.R, R/stepHandleHelpers.R
\name{retrieveMainProcess}
\alias{retrieveMainProcess}
\title{retrieveMainProcess
retrieves the main process data frame of a step by handle}
\usage{
retrieveMainProcess(stepHandle)

retrieveMainProcess(stepHandle)
}
\arguments{
\item{stepHandle}{id of the prepared step}
}
\description{
retrieveMainProcess
retrieves the main process data frame of a step by handle

retrieveMainProcess
retrieves the main process data frame of a step by handle
}
