% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandle.R
\name{retrieveStep}
\alias{retrieveStep}
\title{retrieves complete step description}
\usage{
retrieveStep(stepHandle)
}
\arguments{
\item{stepHandle}{id of the prepared step}
}
\description{
retrieves complete step description
}
\references{
ics1140
}
