% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{retrieveWorkflow}
\alias{retrieveWorkflow}
\title{retrieves complete workflow description, can be used to refresh a workflow}
\usage{
retrieveWorkflow(workflowHandle)
}
\arguments{
\item{workflowHandle}{handle of the workflow}
}
\description{
retrieves complete workflow description, can be used to refresh a workflow
}
