% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStep.R
\name{runStep}
\alias{runStep}
\title{runStep}
\usage{
runStep(handle)
}
\arguments{
\item{handle}{check if a step with the same configuration already exists in the tree}
}
\description{
runStep
}
\references{
ics1140
}
