% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_realise.R
\name{runStep_deprecated}
\alias{runStep_deprecated}
\title{Run step using deprecated approach}
\usage{
runStep_deprecated(step)
}
\description{
Run step using deprecated approach
}
\keyword{internal}
