% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGridArguments.R
\name{setGridArgument}
\alias{setGridArgument}
\title{setGridArgument}
\usage{
setGridArgument(processId, argumentName, argumentValue, update = T)
}
\arguments{
\item{processId}{id of the preocess}

\item{argumentName}{name of the grid argument}

\item{argumentValue}{value of the grid argument}

\item{update}{updates if one value is already set, if not adds, defaults to TRUE}
}
\description{
setGridArgument
}
\references{
ics1222
}
