% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{setProcessVariables}
\alias{setProcessVariables}
\title{setProcessVariables}
\usage{
setProcessVariables(
  stepId,
  processId,
  processType = "main",
  name = "Main",
  selected = TRUE,
  runserverId = "",
  toolId = "",
  runserverToolId = "",
  toolIgnorePatterns = "",
  mainProcess = T,
  gridTool,
  toolArguments = NULL
)
}
\arguments{
\item{stepId}{id of the step the variables refer to}

\item{processId}{id of the process the variables refer to}

\item{processType}{processType}

\item{name}{name}

\item{selected}{selected}

\item{runserverId}{runserverId}

\item{toolId}{toolId}

\item{runserverToolId}{runserverToolId}

\item{toolIgnorePatterns}{toolIgnorePatterns}

\item{mainProcess}{if this process is status relevant, there needs to be exactly one mainProcess}

\item{gridTool}{gridTool}

\item{toolArguments}{toolArguments}
}
\description{
setProcessVariables
}
