% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedTokenManager.R
\name{setSharedSecret}
\alias{setSharedSecret}
\title{Set shared secret for current session}
\usage{
setSharedSecret(secret)
}
\arguments{
\item{secret}{The shared secret UUID}
}
\description{
Sets the shared secret for the current session
}
