% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setStepBreakpoint}
\alias{setStepBreakpoint}
\title{sets a step breakpoint, the step run is not started even if realiseStep is called with run=T}
\usage{
setStepBreakpoint(stepHandle, breakpoint = T)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{breakpoint}{boolean, true inhibits step execution in workflows/realiseStep}
}
\description{
sets a step breakpoint, the step run is not started even if realiseStep is called with run=T
}
