% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setStepDescription}
\alias{setStepDescription}
\title{sets the description of the step}
\usage{
setStepDescription(stepHandle, description)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{description}{description of the step}
}
\description{
sets the description of the step
}
