% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setStepFinishCondition}
\alias{setStepFinishCondition}
\title{sets the finish runserver and runserver tool by name, finishStep only stops if the step is executed and finished with this combination}
\usage{
setStepFinishCondition(stepHandle, runserverName, runserverTool)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{runserverName}{name of the runserver}

\item{runserverTool}{name of the runserverTool}
}
\description{
sets the finish runserver and runserver tool by name, finishStep only stops if the step is executed and finished with this combination
}
