% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setStepName}
\alias{setStepName}
\title{sets the name of the step}
\usage{
setStepName(stepHandle, stepName)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{stepName}{name of the step}
}
\description{
sets the name of the step
}
