% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setStepParent}
\alias{setStepParent}
\title{sets the parent step}
\usage{
setStepParent(stepHandle, parentIdent, inheritFromParent = F)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{parentIdent}{ident of the containing tree}

\item{inheritFromParent}{if input files and settings should be taken over from the parent}
}
\description{
sets the parent step
}
