% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setStepRationale}
\alias{setStepRationale}
\title{sets the rationale of the step}
\usage{
setStepRationale(stepHandle, rationale)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{rationale}{rationale of the step}
}
\description{
sets the rationale of the step
}
