% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setStepReuse}
\alias{setStepReuse}
\title{sets a step reusage, the step is not created if an identical finished step already exists in the tree}
\usage{
setStepReuse(stepHandle, reuse = T)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{reuse}{boolean, true inhibits recreation of identical in workflows/realiseStep}
}
\description{
sets a step reusage, the step is not created if an identical finished step already exists in the tree
}
