% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setStepRunserverToolName}
\alias{setStepRunserverToolName}
\title{sets the runserver tool by name}
\usage{
setStepRunserverToolName(stepHandle, runserverToolName)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{runserverToolName}{name of the runserver tool}
}
\description{
sets the runserver tool by name
}
