% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setStepToolName}
\alias{setStepToolName}
\title{sets the tool by name}
\usage{
setStepToolName(stepHandle, toolName)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{toolName}{name of the  tool}
}
\description{
sets the tool by name
}
