% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setStepTree}
\alias{setStepTree}
\title{sets the containing tree}
\usage{
setStepTree(stepHandle, treeIdent)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{treeIdent}{ident of the containing tree}
}
\description{
sets the containing tree
}
