% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setStepWorkflow}
\alias{setStepWorkflow}
\title{sets the improveR workflow}
\usage{
setStepWorkflow(stepHandle, workflowHandle)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{workflowHandle}{handle of the workflow}
}
\description{
sets the improveR workflow
}
