% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setWorkflowTreeIdent}
\alias{setWorkflowTreeIdent}
\title{setWorkflowTreeIdent
sets the ident for the existing tree that the steps are created in.
Use detachWorkflowFromTrees to remove treeIdent again.}
\usage{
setWorkflowTreeIdent(workflowHandle, treeIdent)
}
\arguments{
\item{workflowHandle}{handle of the workflow}

\item{treeIdent}{ident of the tree}
}
\description{
setWorkflowTreeIdent
sets the ident for the existing tree that the steps are created in.
Use detachWorkflowFromTrees to remove treeIdent again.
}
