% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setWorkflowTreeName}
\alias{setWorkflowTreeName}
\title{setWorkflowTreeName
sets the name for the tree that the steps are created in. Only used if no treeIdent is set.
Use detachWorkflowFromTrees to remove treeIdent}
\usage{
setWorkflowTreeName(workflowHandle, treeName)
}
\arguments{
\item{workflowHandle}{handle of the workflow}

\item{treeName}{name of the tree}
}
\description{
setWorkflowTreeName
sets the name for the tree that the steps are created in. Only used if no treeIdent is set.
Use detachWorkflowFromTrees to remove treeIdent
}
