% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setWorkflowTreeRootFolder}
\alias{setWorkflowTreeRootFolder}
\title{setWorkflowTreeRootFolder
sets the path to a root folder for the tree that the steps are created in. Only used if no treeIdent is set.
Use detachWorkflowFromTrees to remove treeIdent}
\usage{
setWorkflowTreeRootFolder(workflowHandle, rootFolder)
}
\arguments{
\item{workflowHandle}{handle of the workflow}

\item{rootFolder}{path to the rootFolder}
}
\description{
setWorkflowTreeRootFolder
sets the path to a root folder for the tree that the steps are created in. Only used if no treeIdent is set.
Use detachWorkflowFromTrees to remove treeIdent
}
