% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHTML.R
\name{showHTML}
\alias{showHTML}
\title{Includes HTML in rMarkdown, creates a string that has to be outputted}
\usage{
showHTML(
  ident,
  from = pwd(),
  addAsLink = TRUE,
  caption = "",
  includeCaption = T
)
}
\arguments{
\item{ident}{path, resource or entity ID of the picture}

\item{from}{used for relative paths, by default pwd is used, which is initiated with the step that started improveR}

\item{addAsLink}{creates a link in inventory if TRUE, warn improveClean has to be run at the end}

\item{caption}{by default entityID and lastmodified are the caption, here alternative text can be provided}

\item{includeCaption}{display the caption}
}
\description{
Includes HTML in rMarkdown, creates a string that has to be outputted
}
\references{
ics1141
}
