% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getR.R
\name{sourceR}
\alias{sourceR}
\title{sources R objects}
\usage{
sourceR(ident, from = pwd(), addAsLink = TRUE)
}
\arguments{
\item{ident}{path, resource or entity ID of the R object}

\item{from}{used for relative paths, by default pwd is used, which is initiated with the step that started improveR}

\item{addAsLink}{creates a link in inventory if TRUE, warn improveClean has to be run at the end}
}
\description{
sources R objects
}
\references{
ics1141
}
