% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericLoadResourceEntities.R
\name{strip}
\alias{strip}
\title{strip removes the information dataframe around subentities like meta data, history, children or audittrail}
\usage{
strip(data)
}
\arguments{
\item{data}{a data frame with a nested frame in $data}
}
\description{
strip removes the information dataframe around subentities like meta data, history, children or audittrail
}
