% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliNew.R
\name{toolsCli}
\alias{toolsCli}
\title{List Available Tools}
\usage{
toolsCli()
}
\value{
CLI output with tool listing.
}
\description{
Lists tool categories and tools available in improve. Useful for
determining the \code{toolCategory} and \code{tool} arguments for
\code{\link{createStepCli}}.
}
