% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFullChildResources.R
\name{unloadFullChildResources}
\alias{unloadFullChildResources}
\title{unloadFullChildResources}
\usage{
unloadFullChildResources(ident)
}
\arguments{
\item{ident}{id}
}
\description{
unloadFullChildResources
}
\references{
ics1085
}
