% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadparentStep.R
\name{unloadParentStep}
\alias{unloadParentStep}
\title{Unload Parent Step}
\usage{
unloadParentStep(ident, from = pwd())
}
\arguments{
\item{ident}{resourceID, entityId or path to the step.}

\item{from}{path working directory, default is the calling step}
}
\description{
Unload Parent Step
}
\references{
ics1209
}
