% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProcess.R
\name{unloadProcessesForStep}
\alias{unloadProcessesForStep}
\title{unloadProcessesForStep}
\usage{
unloadProcessesForStep(stepIdent)
}
\arguments{
\item{stepIdent}{resourceId of the step or the step}
}
\description{
unloadProcessesForStep
}
\references{
ics1218
}
