% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProcess.R
\name{unloadProcessRuns}
\alias{unloadProcessRuns}
\title{unloadProcessRuns}
\usage{
unloadProcessRuns(processId)
}
\arguments{
\item{processId}{id of the process}
}
\description{
unloadProcessRuns
}
\references{
ics1218
}
