% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProcess.R
\name{unloadProcessVariables}
\alias{unloadProcessVariables}
\title{unloadProcessVariables}
\usage{
unloadProcessVariables(processId)
}
\arguments{
\item{processId}{id of the process}
}
\description{
unloadProcessVariables
}
\references{
ics1218
}
