% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceRelations.R
\name{unloadRelationTypes}
\alias{unloadRelationTypes}
\title{Unloads All Relation Types}
\usage{
unloadRelationTypes(conn)
}
\arguments{
\item{conn}{database connection}
}
\description{
Unloads All Relation Types
}
