% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceRelations.R
\name{unloadResourceRelations}
\alias{unloadResourceRelations}
\title{Unloads All Resource Relations}
\usage{
unloadResourceRelations(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
Unloads All Resource Relations
}
\references{
ics1044
}
