% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{unloadToolsForCategory}
\alias{unloadToolsForCategory}
\title{unloadToolsForCategory}
\usage{
unloadToolsForCategory(categoryId)
}
\arguments{
\item{categoryId}{categoryId of the tools}
}
\description{
unloadToolsForCategory
}
\references{
ics1230
}
