% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliSetup.R
\name{unpackOsVersion}
\alias{unpackOsVersion}
\title{Unpack OS Version}
\usage{
unpackOsVersion(osKey, possibleFiles)
}
\arguments{
\item{osKey}{A string representing the OS key (e.g., "win32", "lnx", "macosx").}

\item{possibleFiles}{a list of zip files}
}
\description{
Unpacks the CLI files for the specified OS key.
}
