% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFullChildResources.R
\name{updateFullChildResources}
\alias{updateFullChildResources}
\title{Update Full Child Resources}
\usage{
updateFullChildResources(ident)
}
\arguments{
\item{ident}{id}
}
\description{
Update Full Child Resources
}
\references{
ics1085
}
