% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGridArguments.R
\name{updateGridArgument}
\alias{updateGridArgument}
\title{updateGridArgument}
\usage{
updateGridArgument(processId, argumentName, argumentValue)
}
\arguments{
\item{processId}{id of the preocess}

\item{argumentName}{name of the grid argument}

\item{argumentValue}{value of the grid argument}
}
\description{
updateGridArgument
}
\references{
ics1222
}
