% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGridArguments.R
\name{updateGridArguments}
\alias{updateGridArguments}
\title{updateGridArguments reloads the grid arguments from the repository}
\usage{
updateGridArguments(gridProvider)
}
\arguments{
\item{gridProvider}{label of the grid provider}
}
\description{
updateGridArguments reloads the grid arguments from the repository
}
\references{
ics1216
}
