% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateLinks.R
\name{updateLinks}
\alias{updateLinks}
\title{Update link resource(s)}
\usage{
updateLinks(links, comment = "update outdated")
}
\arguments{
\item{links}{Vector of resource IDs or a data.frame with a column 'entityId'}

\item{comment}{Optional comment for the update}
}
\description{
Updates one or more link resources by resourceId.
Only updates resources of type "Link".
After updating, unloads the parent resource and its children to clear cache.
}
