% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadparentStep.R
\name{updateParentStep}
\alias{updateParentStep}
\title{updateParentStep reloads the runservers from the repository}
\usage{
updateParentStep(ident, from = pwd())
}
\arguments{
\item{ident}{resourceID, entityId or path to the step.}

\item{from}{path working directory, default is the calling step}
}
\description{
updateParentStep reloads the runservers from the repository
}
\references{
ics1209
}
