% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProcess.R
\name{updateProcessesForStep}
\alias{updateProcessesForStep}
\title{updateProcessesForStep reloads the processes for a step}
\usage{
updateProcessesForStep(stepIdent)
}
\arguments{
\item{stepIdent}{resourceId of the step or the step}
}
\description{
updateProcessesForStep reloads the processes for a step
}
\references{
ics1218
}
