% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProcess.R
\name{updateProcessGridArguments}
\alias{updateProcessGridArguments}
\title{updateLoadProcessGridArguments reloads the process grid arguments for a process}
\usage{
updateProcessGridArguments(processId)
}
\arguments{
\item{processId}{id of the process}
}
\description{
updateLoadProcessGridArguments reloads the process grid arguments for a process
}
\references{
ics1218
}
