% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProcess.R
\name{updateProcessRuns}
\alias{updateProcessRuns}
\title{updateProcessRuns reloads the process runs for a process}
\usage{
updateProcessRuns(processId)
}
\arguments{
\item{processId}{id of the process}
}
\description{
updateProcessRuns reloads the process runs for a process
}
\references{
ics1218
}
