% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProcess.R
\name{updateProcessVariables}
\alias{updateProcessVariables}
\title{updateProcessVariables reloads the process variables for a process}
\usage{
updateProcessVariables(processId)
}
\arguments{
\item{processId}{id of the process}
}
\description{
updateProcessVariables reloads the process variables for a process
}
\references{
ics1218
}
