% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReferences.R
\name{updateReferences}
\alias{updateReferences}
\title{Update References}
\usage{
updateReferences(ident)
}
\arguments{
\item{ident}{id}
}
\description{
Update References
}
\references{
ics1206
}
