% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceRelations.R
\name{updateRelationTypes}
\alias{updateRelationTypes}
\title{Reloads the Relation Types}
\usage{
updateRelationTypes(conn)
}
\arguments{
\item{conn}{database connection}
}
\description{
Reloads the Relation Types
}
