% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResource.R
\name{updateResource}
\alias{updateResource}
\title{Update Resource}
\usage{
updateResource(ident, from = pwd())
}
\arguments{
\item{ident}{id}

\item{from}{pwd for relative path}
}
\description{
Updates a resource.
}
\references{
ics1090
}
