% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{updateReviewComments}
\alias{updateReviewComments}
\title{Reloads the Review Comments for a Review}
\usage{
updateReviewComments(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
Reloads the Review Comments for a Review
}
\references{
ics1208
}
