% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReview.R
\name{updateReviewEntryComments}
\alias{updateReviewEntryComments}
\title{Reloads the Comments for a Review Entry}
\usage{
updateReviewEntryComments(ident, entryId, from = pwd())
}
\arguments{
\item{ident}{Identifier of the review. Can be a path, resource ID, entity ID,
or a data frame row from \code{loadResource()}.}

\item{entryId}{Character. ID (UUID) of the review entry.}

\item{from}{Base path for resolving relative paths. Defaults to \code{pwd()}.}
}
\value{
A data frame of review entry comments, or \code{NULL} if none exist.
}
\description{
Reloads the Comments for a Review Entry
}
\references{
ics1543
}
