% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReview.R
\name{updateReviews}
\alias{updateReviews}
\title{Reloads the Reviews}
\usage{
updateReviews()
}
\value{
A data frame of reviews, or \code{NULL} if none exist.
}
\description{
Reloads the Reviews
}
\references{
ics348
}
