% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepConfigData.R
\name{updateRunservers}
\alias{updateRunservers}
\title{updateRunservers reloads the runservers from the repository}
\usage{
updateRunservers()
}
\description{
updateRunservers reloads the runservers from the repository
}
\references{
ics1226
}
