% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{updateToolCategories}
\alias{updateToolCategories}
\title{updateToolCategories reloads the tool categories from the repository}
\usage{
updateToolCategories()
}
\description{
updateToolCategories reloads the tool categories from the repository
}
\references{
ics1229
}
