% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{updateToolsForCategory}
\alias{updateToolsForCategory}
\title{updateToolsForCategory reloads the category tools from the repository}
\usage{
updateToolsForCategory(categoryId)
}
\arguments{
\item{categoryId}{categoryId of the category}
}
\description{
updateToolsForCategory reloads the category tools from the repository
}
\references{
ics1230
}
