% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepConfigData.R
\name{updateToolsForRunserver}
\alias{updateToolsForRunserver}
\title{updateToolsForRunserver reloads the runserver tools from the repository}
\usage{
updateToolsForRunserver(runserverId)
}
\arguments{
\item{runserverId}{resourceId of the runserver}
}
\description{
updateToolsForRunserver reloads the runserver tools from the repository
}
\references{
ics1227
}
