% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worflowOrder.R
\name{usage}
\alias{usage}
\title{retrieves a data frame holding the usage of a step within the given workflow.
One workflow can span multiple analsis trees
random order returned}
\usage{
usage(workflow, stepHandle)
}
\arguments{
\item{workflow}{as df}

\item{stepHandle}{id of the lineage step}
}
\description{
retrieves a data frame holding the usage of a step within the given workflow.
One workflow can span multiple analsis trees
random order returned
}
\references{
ics1231
}
