% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers.R
\name{validateDuplicateReviewer}
\alias{validateDuplicateReviewer}
\title{helper function to check if another reviewer with the given user already exists for the review}
\usage{
validateDuplicateReviewer(reviewId, userId, username)
}
\arguments{
\item{reviewId}{id (UUID) of the review}

\item{userId}{id (UUID) of the user}

\item{username}{name of the user}
}
\description{
helper function to check if another reviewer with the given user already exists for the review
}
