% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewEntryComments.R
\name{validateEntry}
\alias{validateEntry}
\title{helper function to validate the review entry's existence}
\usage{
validateEntry(reviewId, entryId)
}
\arguments{
\item{reviewId}{id (UUID) of the review}

\item{entryId}{id (UUID) of the review entry whose existence is to be checked}
}
\description{
helper function to validate the review entry's existence
}
