% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateHelper.R
\name{validateParams}
\alias{validateParams}
\title{helper function to check if any parameter is NA or NULL}
\usage{
validateParams(params)
}
\arguments{
\item{params}{list of parameters that are to be checked}
}
\description{
helper function to check if any parameter is NA or NULL
}
