% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateHelper.R
\name{validateRelationType}
\alias{validateRelationType}
\title{helper function to validate relation type existence}
\usage{
validateRelationType(relationTypeId, conn = NULL)
}
\arguments{
\item{relationTypeId}{id (UUID) of the relation type whose existence is to be checked}

\item{conn}{database connection}
}
\description{
helper function to validate relation type existence
}
