% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceRelations.R, R/reviews.R
\name{validateResource}
\alias{validateResource}
\title{helper function to validate resource existence}
\usage{
validateResource(resourceId)

validateResource(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource whose existence is to be checked}
}
\description{
helper function to validate resource existence

helper function to validate the resource's existence
}
