% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceRelations.R
\name{validateResourceRelation}
\alias{validateResourceRelation}
\title{helper function to validate relation existence}
\usage{
validateResourceRelation(resourceId, relationId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}

\item{relationId}{id (UUID) of the relation}
}
\description{
helper function to validate relation existence
}
