% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers.R
\name{validateReview}
\alias{validateReview}
\title{helper function to validate the review's existence}
\usage{
validateReview(reviewId)
}
\arguments{
\item{reviewId}{id (UUID) of the review whose existence is to be checked}
}
\description{
helper function to validate the review's existence
}
