% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers.R
\name{validateReviewer}
\alias{validateReviewer}
\title{helper function to validate the reviewer's existence}
\usage{
validateReviewer(reviewId, reviewerId)
}
\arguments{
\item{reviewId}{id (UUID) of the review}

\item{reviewerId}{id (UUID) of the reviewer whose existence is to be checked}
}
\description{
helper function to validate the reviewer's existence
}
