% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewComments.R
\name{validateReviewState}
\alias{validateReviewState}
\title{helper function to validate the state of the review}
\usage{
validateReviewState(reviewId)
}
\arguments{
\item{reviewId}{id (UUID) of the review}
}
\description{
helper function to validate the state of the review
}
