% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceRelations.R
\name{validateTargetResourceIdDuplicate}
\alias{validateTargetResourceIdDuplicate}
\title{helper function to check if another relation with the given targetResourceId already exists for the resource}
\usage{
validateTargetResourceIdDuplicate(resourceId, targetResourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}

\item{targetResourceId}{id (UUID) of the target resource}
}
\description{
helper function to check if another relation with the given targetResourceId already exists for the resource
}
