% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers.R
\name{validateUser}
\alias{validateUser}
\title{helper function to validate the user's existence}
\usage{
validateUser(userId, username)
}
\arguments{
\item{userId}{id (UUID) of the user whose existence is to be checked}

\item{username}{name of the user whose existence is to be checked}
}
\description{
helper function to validate the user's existence
}
