% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoami.R
\name{whoami}
\alias{whoami}
\title{retrieves the name of the user that started the step or was used to authenticate}
\usage{
whoami()
}
\description{
retrieves the name of the user that started the step or was used to authenticate
}
\references{
ics1251
}
