% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xor.R
\name{xorDecrypt}
\alias{xorDecrypt}
\title{decrypts a string with a one time pad}
\usage{
xorDecrypt(value, key)
}
\arguments{
\item{value}{the value to decrypt}

\item{key}{the key}
}
\description{
decryption using xor
}
