% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xor.R
\name{xorEncrypt}
\alias{xorEncrypt}
\title{encrypts a string with a one time pad}
\usage{
xorEncrypt(value, key)
}
\arguments{
\item{value}{the value to encrypt}

\item{key}{the key}
}
\description{
encryption using xor
}
