% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{addBulkMetaDate}
\alias{addBulkMetaDate}
\title{Adds Metadata Value To A Resource, Always Adds, Never Updates}
\usage{
addBulkMetaDate(ident, descriptorNameValueList, scope = "Improve Client")
}
\arguments{
\item{ident}{the resource id of the resource this metadata value is attached to}

\item{descriptorNameValueList}{a list of lists, containing descriptorName and value: this descriptorName has to exist, check in preferences, the value has to match the type of the descriptor. For dates, posix date has to be used, for LOVs the text has to exist}

\item{scope}{the scope for metadata, default "Improve Client"}
}
\description{
Adds Metadata Value To A Resource, Always Adds, Never Updates
}
\references{
ics1137
}
