% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{addGroupUser}
\alias{addGroupUser}
\title{Add a User to a Group}
\usage{
addGroupUser(groupId, userId)
}
\arguments{
\item{groupId}{Character. The ID of the group.}

\item{userId}{Character. The ID of the user to add.}
}
\value{
A list of the user's group memberships after the addition,
or \code{NULL} on failure.
}
\description{
Adds an existing user to a group by creating a group membership.
}
\examples{
\dontrun{
addGroupUser("group-id", "user-id")
}
}
\seealso{
\code{\link{removeGroupUser}}, \code{\link{loadGroupUsers}}, \code{\link{createGroup}}
}
