% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{addMetaDate}
\alias{addMetaDate}
\title{Adds Metadata Value To A Resource, Always Adds, Never Updates}
\usage{
addMetaDate(ident, descriptorName, value, scope = "Improve Client")
}
\arguments{
\item{ident}{the resource id of the resource this metadata value is attached to}

\item{descriptorName}{this descriptorName has to exist, check in preferences}

\item{value}{the value has to match the type of the descriptor. For dates, posix date has to be used, for LOVs the text has to exist}

\item{scope}{the scope for metadata, default "Improve Client"}
}
\description{
Adds Metadata Value To A Resource, Always Adds, Never Updates
}
\references{
ics1137
}
