% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{addStepGridArgument}
\alias{addStepGridArgument}
\title{adds a new grid argument to the step}
\usage{
addStepGridArgument(stepHandle, argumentName, argumentValue, process = "Main")
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{argumentName}{the ident of the file in the repository}

\item{argumentValue}{leave empty if you want to use the same name as the used file}

\item{process}{the name of the process, default = Main. If the process does not yet exist it is created}
}
\description{
adds a new grid argument to the step
}
