% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{addStepRemoteFile}
\alias{addStepRemoteFile}
\title{adds a new remote file to the step}
\usage{
addStepRemoteFile(
  stepHandle,
  ident = NULL,
  name = NULL,
  asLink = T,
  variableName = NULL,
  sourceHandle = NULL,
  sourceName = NULL
)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{ident}{the ident of the file in the repository}

\item{name}{leave empty if you want to use the same name as the used file}

\item{asLink, }{if file should be copied or linked, boolean, defaults to link}

\item{variableName}{the name of the variable the file should be bound to, optional}

\item{sourceHandle}{stepHandle if ident is relative to another step in the workflow}

\item{sourceName}{name if ident is relative to another step in the workflow}
}
\description{
adds a new remote file to the step
}
