% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{attachStep}
\alias{attachStep}
\title{attaches a step to its parent}
\usage{
attachStep(ident, parent, from = pwd())
}
\arguments{
\item{ident}{ident of the step}

\item{parent}{ident of the new parent}

\item{from}{from if a relative path is used}
}
\description{
attaches a step to its parent
}
\references{
ics1225
}
