% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{byNotEmpty}
\alias{byNotEmpty}
\title{byNotEmpty by that automatically uses the complete dataframe and returns null if an empty or non existing data frame is handed over}
\usage{
byNotEmpty(df, func)
}
\arguments{
\item{df}{the data frame}

\item{func}{the function that is applied to each row of the data frame}
}
\description{
byNotEmpty by that automatically uses the complete dataframe and returns null if an empty or non existing data frame is handed over
}
