% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareStep.R
\name{changeStepDescription}
\alias{changeStepDescription}
\title{changes the step description}
\usage{
changeStepDescription(ident, from = pwd(), description)
}
\arguments{
\item{ident}{ident of the step}

\item{from}{from if a relative path is used}

\item{description}{new step desciption}
}
\description{
changes the step description
}
\references{
ics1217
}
