% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareStep.R
\name{changeStepRationale}
\alias{changeStepRationale}
\title{changes the step rationale}
\usage{
changeStepRationale(ident, from = pwd(), rationale)
}
\arguments{
\item{ident}{ident of the step}

\item{from}{from if a relative path is used}

\item{rationale}{new step rationale}
}
\description{
changes the step rationale
}
\references{
ics1217
}
