% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveConnect.R
\name{checkConnect}
\alias{checkConnect}
\title{Check Connect}
\usage{
checkConnect()
}
\value{
invisible TRUE if connection is valid, otherwise attempts reconnection
}
\description{
Checks if the current connection is still valid by attempting to load the IMPROVER_STEP resource.
If the connection is invalid, it clears connection data and attempts to reconnect.
}
