% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveConnect.R
\name{clearConnectionData}
\alias{clearConnectionData}
\title{clearConnectionData}
\usage{
clearConnectionData()
}
\description{
deletes repoUrl, stepId, and token from the environment variables and calls improveDisconnect
}
\seealso{
\code{\link[=improveDisconnect]{improveDisconnect()}}
}
