% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliClone.R
\name{cloneCli}
\alias{cloneCli}
\title{Clone CLI Resource}
\usage{
cloneCli(ident, localPath)
}
\arguments{
\item{ident}{The identifier of the resource to clone.}

\item{localPath}{The local path where the resource will be cloned.}
}
\description{
Clones a resource from the CLI.
}
