% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertImproveTimestampToPosix.R
\name{convertImproveTimestampToPosix}
\alias{convertImproveTimestampToPosix}
\title{takes a timestamp as used in improve and converts it to a POSIX R time}
\usage{
convertImproveTimestampToPosix(timestamp)
}
\arguments{
\item{timestamp}{as character, numeric or list}
}
\description{
takes a timestamp as used in improve and converts it to a POSIX R time
}
