% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertImproveTimestampToPosix.R
\name{convertImproveTimestampToPosix}
\alias{convertImproveTimestampToPosix}
\title{convertImproveTimestampToPosix}
\usage{
convertImproveTimestampToPosix(timestamp)
}
\arguments{
\item{timestamp}{a character, numeric string or list}
}
\description{
Takes a timestamp as used in improve and converts it to a POSIX R time
}
