% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createAnalysisTree}
\alias{createAnalysisTree}
\title{Create Analysis Tree in improve Repository}
\usage{
createAnalysisTree(targetIdent, treeName = "", comment = "Created by improveR")
}
\arguments{
\item{targetIdent}{Identifier of the parent folder where the analysis tree
should be created.}

\item{treeName}{Character. Name of the new analysis tree.}

\item{comment}{Character. Comment for the creation audit entry. Defaults to
"Created by improveR".}
}
\value{
A list containing the resource object of the created analysis tree.
Returns \code{NULL} on failure.
}
\description{
Creates an Analysis Tree, which is a specialized folder structure for organizing
pharmaceutical analysis workflows.
}
\references{
ics1103
}
\seealso{
\code{\link{createFolder}}, \code{\link{createWorkflow}}
}
