% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createAnalysisTree}
\alias{createAnalysisTree}
\title{creates a workflow/analysis tree in the repository and returns the resource object}
\usage{
createAnalysisTree(
  targetIdent,
  treeName = "",
  comment = "Created by improveRW"
)
}
\arguments{
\item{targetIdent}{one or many target folders, as resource, id or path}

\item{treeName}{one or many treeNames}

\item{comment}{defaults to Created by improveRW}
}
\description{
creates a workflow/analysis tree in the repository and returns the resource object
}
\references{
ics1103
}
