% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createAnalysisTree}
\alias{createAnalysisTree}
\title{Create an Analysis Tree
Creates an analysis tree in a specified folder and assigns it a specified name.}
\usage{
createAnalysisTree(targetIdent, treeName = "", comment = "Created by improveR")
}
\arguments{
\item{targetIdent}{ident of the folder where the analysis tree is to be created in.}

\item{treeName}{name of the newly created analysis tree}

\item{comment}{defaults to "Created by improveR"}
}
\description{
Create an Analysis Tree
Creates an analysis tree in a specified folder and assigns it a specified name.
}
\references{
ics1103
}
