% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createExternalLink}
\alias{createExternalLink}
\title{Create an External Link}
\usage{
createExternalLink(
  targetIdent,
  linkName = "",
  url,
  comment = "Created by improveR"
)
}
\arguments{
\item{targetIdent}{ident of a folder, analysis tree, or step where the external link should be created}

\item{linkName}{Name of the link}

\item{url}{url pointing at the the external resource, e.g., a database.}

\item{comment}{defaults to Created by improveR}
}
\description{
An externa link is a link pointing to a resource outside of the respository.
The function createExternalLink() establishes such a link within a folder, analysis tree, or step.
}
\references{
ics1104
}
