% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favorites.R
\name{createFavoriteFolder}
\alias{createFavoriteFolder}
\title{Create Favorite Folder}
\usage{
createFavoriteFolder(name, parentId = NULL, comment = "")
}
\arguments{
\item{name}{Name of the new favorites folder.}

\item{parentId}{Optional resource ID of the parent favorites folder. If
\code{NULL}, the folder is created at the top level.}

\item{comment}{Optional comment for the operation.}
}
\value{
A data frame with the created favorites folder details, or \code{NULL} on failure.
}
\description{
Creates a new folder inside the user's favorites collection for organising
favorite resources.
}
\references{
ics1805, ics1806
}
