% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createFile}
\alias{createFile}
\title{Create a File}
\usage{
createFile(
  targetIdent,
  fileName = "",
  localPath = "",
  comment = "Created by improveR"
)
}
\arguments{
\item{targetIdent}{one or many target folders, as resource, id or path}

\item{fileName}{Name of the file to be created.}

\item{localPath}{path to a local file for initial file contents}

\item{comment}{defaults to "Created by improveR"}
}
\description{
The function \code{createFile()} generates a file in the repository and returns the resource object.
}
\references{
ics1102
}
