% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createFolder}
\alias{createFolder}
\title{Create Folder in improve Repository}
\usage{
createFolder(targetIdent, folderName = "", comment = "Created by improveR")
}
\arguments{
\item{targetIdent}{Identifier(s) of the parent location where the new folder(s)
should be created. Can be a path, resource ID, or entity ID.}

\item{folderName}{Character. Name(s) of the new folder(s).}

\item{comment}{Character. Comment for the creation audit entry. Defaults to
"Created by improveR".}
}
\value{
A list containing the resource object(s) of the created folder(s).
Each resource object contains metadata like \code{resourceId}, \code{entityId}, \code{path}, etc.
Returns \code{NULL} on failure.
}
\description{
Creates one or multiple folders in the repository at the specified location.
}
\examples{
\dontrun{
# Create a folder in the repository root
createFolder(targetIdent = "/", folderName = "folderInRoot")

# Create a folder inside an existing folder by path
createFolder(
  targetIdent = "/improve-tutorial/",
  folderName = "folderInImproveTutorial"
)
}
}
\references{
ics1101
}
\seealso{
\code{\link{createFile}}, \code{\link{createAnalysisTree}}
}
