% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createFolder}
\alias{createFolder}
\title{creates a folder in the repository and returns the resource object}
\usage{
createFolder(targetIdent, folderName = "", comment = "Created by improveRW")
}
\arguments{
\item{targetIdent}{one or many target folders, as resource, id or path}

\item{folderName}{one or many folderNames}

\item{comment}{defaults to Created by improveRW}
}
\description{
creates a folder in the repository and returns the resource object
}
\references{
ics1101
}
