% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{createLink}
\alias{createLink}
\title{createLink
creates a link to a resource, currently always the latest version of the resource is used}
\usage{
createLink(linkContainer, links, linkName = "")
}
\arguments{
\item{linkContainer}{where to create the link}

\item{links}{to which resource to link}

\item{linkName}{name for the link, if this is not given the name of the linked resource is used. Can only be used if only one resource is linked}
}
\description{
createLink
creates a link to a resource, currently always the latest version of the resource is used
}
\references{
ics1138
}
