% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{createReview}
\alias{createReview}
\title{creates a new review}
\usage{
createReview(
  name,
  parentPath,
  comment,
  templateId,
  resourceIds,
  reviewerIds,
  dueDate
)
}
\arguments{
\item{name}{name of the review}

\item{parentPath}{path of the parent}

\item{comment}{review comment}

\item{templateId}{id (UUID) of the template}

\item{resourceIds}{vector of resourceIds}

\item{reviewerIds}{vector of reviewerIds}

\item{dueDate}{format: yyyy-mm-dd}
}
\description{
creates a new review
}
\references{
ics1527
}
