% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewComments.R
\name{createReviewComment}
\alias{createReviewComment}
\title{creates a comment for a review}
\usage{
createReviewComment(reviewId, resourceId, comment, commentType)
}
\arguments{
\item{reviewId}{id (UUID) of the review}

\item{resourceId}{id (UUID) of the resource}

\item{comment}{comment}

\item{commentType}{type of the comment}
}
\description{
creates a comment for a review
}
\references{
ics1536
}
