% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewEntryComments.R
\name{createReviewEntryComment}
\alias{createReviewEntryComment}
\title{Creates a Comment for a Review Entry}
\usage{
createReviewEntryComment(reviewId, entryId, comment)
}
\arguments{
\item{reviewId}{id (UUID) of the review}

\item{entryId}{id (UUID) of the resource}

\item{comment}{comment}
}
\description{
Creates a Comment for a Review Entry
}
\references{
ics1544
}
