% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers.R
\name{createReviewer}
\alias{createReviewer}
\title{Adds a User as Reviewer to a Review}
\usage{
createReviewer(reviewId, userId, username)
}
\arguments{
\item{reviewId}{id (UUID) of the review}

\item{userId}{id (UUID) of the user}

\item{username}{name of the user}
}
\description{
Adds a User as Reviewer to a Review
}
\references{
ics368
}
