% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStepEnv.R
\name{createStepEnv}
\alias{createStepEnv}
\title{Create a new step environment}
\usage{
createStepEnv(stepDf = NULL, workflow = NULL)
}
\arguments{
\item{stepDf}{Data frame with step metadata}

\item{workflow}{Workflow environment this step belongs to}
}
\value{
An environment representing the step
}
\description{
Constructs a new step object with encapsulated state, navigation environments, and public API.
}
