% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStepTemplateEnv.R
\name{createStepTemplateEnv}
\alias{createStepTemplateEnv}
\title{Create a new step template environment}
\usage{
createStepTemplateEnv(treeIdent = NULL, stepDf = NULL, workflow = NULL)
}
\arguments{
\item{stepDf}{Data frame with step metadata (optional)}

\item{workflow}{Workflow environment this template belongs to (optional)}
}
\value{
An environment representing the step template
}
\description{
Constructs a new step template object with encapsulated state and public API.
}
