% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWorkflowTemplate.R
\name{createWorkflowTemplateEnv}
\alias{createWorkflowTemplateEnv}
\title{Create a workflow template environment from an existing workflow}
\usage{
createWorkflowTemplateEnv(workflow = NULL, addParental = F)
}
\arguments{
\item{workflow}{The workflow environment to template}

\item{addParental}{if T the created stepTemplates have the steps from the workflow as parents (default: FALSE)}
}
\value{
An environment representing the workflow template
}
\description{
Extracts all steps, relationships, parameters, and files from the workflow
and builds a reusable workflow template environment.
}
