% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepWorkflowCopy.R
\name{deepWorkflowCopy}
\alias{deepWorkflowCopy}
\title{deepWorkflowCopy}
\usage{
deepWorkflowCopy(workflow, targetTree = NULL, createParentalRelation = F)
}
\arguments{
\item{workflow}{a workflow handle or a workflow data.frame, changes to steps in workflow data.frame will be persisted by execution. consider creating a deep copy}

\item{targetTree}{the tree the steps should be executed in, default NULL, if NULL treeIdent of the steps is kept}

\item{createParentalRelation}{create the step as child step of the step in the source workflow, default FALSE}
}
\description{
deepWorkflowCopy
}
\references{
ics1211
}
