% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{delete}
\alias{delete}
\title{deletes resources
multiple sources can be deleted at once}
\usage{
delete(res)
}
\arguments{
\item{res}{the resources to be deleted}
}
\description{
deletes resources
multiple sources can be deleted at once
}
\references{
ics1139
}
