% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{deleteGroup}
\alias{deleteGroup}
\title{Delete a Group}
\usage{
deleteGroup(groupId)
}
\arguments{
\item{groupId}{Character. The ID of the group to delete.}
}
\value{
\code{TRUE} if the group was deleted successfully, \code{FALSE} otherwise.
}
\description{
Deletes an existing group from the improve repository.
}
\examples{
\dontrun{
deleteGroup("some-group-id")
}
}
\seealso{
\code{\link{createGroup}}, \code{\link{loadGroups}}
}
