% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{deleteMetaDate}
\alias{deleteMetaDate}
\title{deletes metadata from a resource}
\usage{
deleteMetaDate(ident, descriptorName)
}
\arguments{
\item{ident}{the resource id of the resource this metadata value is attached to}

\item{descriptorName}{this descriptorName has to exist, check in preferences}
}
\description{
deletes metadata from a resource
}
\references{
ics1137
}
