% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceRelations.R
\name{deleteResourceRelation}
\alias{deleteResourceRelation}
\title{deletes a resource relation}
\usage{
deleteResourceRelation(resourceId, relationId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}

\item{relationId}{id (UUID) of the relation}
}
\description{
deletes a resource relation
}
\references{
ics1044
}
