% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceRelations.R
\name{deleteResourceRelation}
\alias{deleteResourceRelation}
\title{Deletes a Resource Relation}
\usage{
deleteResourceRelation(resourceId, relationId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}

\item{relationId}{id (UUID) of the relation}
}
\description{
Deletes a Resource Relation
}
\references{
ics1044
}
