% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewEntries.R
\name{deleteReviewEntry}
\alias{deleteReviewEntry}
\title{deletes all review entries from a review}
\usage{
deleteReviewEntry(reviewId)
}
\arguments{
\item{reviewId}{id (UUID) of the review}
}
\description{
deletes all review entries from a review
}
\references{
ics365
}
