% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers.R
\name{deleteReviewer}
\alias{deleteReviewer}
\title{removes a reviewer from a review}
\usage{
deleteReviewer(reviewId, reviewerId)
}
\arguments{
\item{reviewId}{id (UUID) of the review}

\item{reviewerId}{id (UUID) of the reviewer}
}
\description{
removes a reviewer from a review
}
\references{
ics369
}
