% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewers.R
\name{deleteReviewer}
\alias{deleteReviewer}
\title{Removes a Reviewer from a Review}
\usage{
deleteReviewer(reviewId, reviewerId)
}
\arguments{
\item{reviewId}{id (UUID) of the review}

\item{reviewerId}{id (UUID) of the reviewer}
}
\description{
Removes a Reviewer from a Review
}
\references{
ics369
}
