% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRun.R
\name{executeFolder}
\alias{executeFolder}
\title{executes a folder with all RMDs and if available before that an init.R and an celanup.R file}
\usage{
executeFolder(path)
}
\arguments{
\item{path}{path to the executable package}
}
\description{
arguments:
}
