% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStep.R
\name{finishRunResource}
\alias{finishRunResource}
\title{finishRunResource}
\usage{
finishRunResource(
  ident,
  from = pwd(),
  runserverName = NULL,
  runserverToolName = NULL
)
}
\arguments{
\item{ident}{the ident to the step}

\item{from}{root for relative pahtes}

\item{runserverName}{if this is set, the step only counts as finished if it was finished with this runserver (needs to be combined with tool), overridden by settings in stephandle}

\item{runserverToolName}{if this is set, the step only counts as finished if it was finished with this tool (needs to be combined with runserver), overridden by settings in stephandle}
}
\description{
finishRunResource
}
\references{
ics1140
}
