% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{getCollaborators}
\alias{getCollaborators}
\title{Get Resource Collaborators}
\usage{
getCollaborators(ident, from = pwd())
}
\arguments{
\item{ident}{Identifier of the resource. Can be a path, resource ID, or entity ID.}

\item{from}{Root path for resolving relative paths. Defaults to \code{pwd()}.}
}
\value{
A data frame of collaborator users. Returns \code{NULL} if the resource
cannot be found. Returns \code{NULL} if there are no collaborators.
}
\description{
Convenience wrapper around \code{\link{getUsersByRole}} that retrieves
users assigned as collaborators of a resource.
}
\examples{
\dontrun{
collabs <- getCollaborators("/Projects/MyProject")
}
}
\seealso{
\code{\link{getOwners}}, \code{\link{getReadOnly}}, \code{\link{getUsersByRole}}
}
