% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{getCopy}
\alias{getCopy}
\title{Get a Local Copy of a File}
\usage{
getCopy(ident, from = pwd(), caption = "", description = "")
}
\arguments{
\item{ident}{Path, resource, or entity ID of the file.}

\item{from}{Used for relative paths. By default, pwd is used (initiated with the step that started improveR).}

\item{caption}{By default, entity ID and last modified are the caption; alternative text can be provided here.}

\item{description}{By default, the filename is the description; alternative text can be provided here.}
}
\description{
Get a Local Copy of a File
}
\references{
ics1141
}
