% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{gets a data table and retrieves the caption and description by the resource ID or entity ID
.xls and .xlsx files assumed to be excel, .rds as RDS, .sas7bdat as SAS all others as CSV}
\usage{
getData(
  ident,
  from = pwd(),
  addAsLink = TRUE,
  caption = "",
  description = "",
  parser = NULL,
  ...
)
}
\arguments{
\item{ident}{path, resource or entity ID of the picture}

\item{from}{used for relative paths, by default pwd is used, which is initiated with the step that started improveR}

\item{addAsLink}{creates a link in inventory if TRUE, warn improveClean has to be run at the end}

\item{caption}{by default entityID and lastmodified are the caption, here alternative text can be provided}

\item{description}{by default filename is the caption, here alternative text can be provided}

\item{parser}{a custom parser can be handed over as argument parser=... . It needs to take a local path as first argument}

\item{...}{pass over arguments for the read function}
}
\description{
gets a data table and retrieves the caption and description by the resource ID or entity ID
.xls and .xlsx files assumed to be excel, .rds as RDS, .sas7bdat as SAS all others as CSV
}
\references{
ics1141
}
