% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{getFile}
\alias{getFile}
\title{gets a file object}
\usage{
getFile(
  ident,
  from = pwd(),
  addAsLink = TRUE,
  caption = "",
  description = "",
  folderName = "data",
  addIdToName = T
)
}
\arguments{
\item{ident}{path, resource or entity ID of the file}

\item{from}{used for relative paths, by default pwd is used, which is initiated with the step that started improveR}

\item{addAsLink}{creates a link in inventory if TRUE, warn improveClean has to be run at the end}

\item{caption}{by default entityID and lastmodified are the caption, here alternative text can be provided}

\item{description}{by default filename is the caption, here alternative text can be provided}

\item{folderName}{by default data the subfolder in the workspace where the file is created}

\item{addIdToName}{by default T prepend the entity ID of the name to avoid collissions}
}
\description{
gets a file object
}
\references{
ics1141
}
