% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{getFile}
\alias{getFile}
\title{Get a File Object}
\usage{
getFile(
  ident,
  from = pwd(),
  addAsLink = TRUE,
  caption = "",
  description = "",
  folderName = "data",
  addIdToName = T,
  refresh = FALSE
)
}
\arguments{
\item{ident}{Path, resource, or entity ID of the file.}

\item{from}{Used for relative paths. By default, pwd is used (initiated with the step that started improveR).}

\item{addAsLink}{Creates a link in the inventory if TRUE; improveClean must be run at the end.}

\item{caption}{By default, entity ID and last modified are the caption; alternative text can be provided here.}

\item{description}{By default, the filename is the description; alternative text can be provided here.}

\item{folderName}{By default, \code{data} is the subfolder in the workspace where the file is created.}

\item{addIdToName}{By default, T prepends the entity ID to the name to avoid collisions.}

\item{refresh}{If TRUE, invalidate cached file content and resource metadata for \code{ident} before fetching. Defaults to FALSE.}
}
\description{
Get a File Object
}
\references{
ics1141
}
