% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{getFilesFromFolder}
\alias{getFilesFromFolder}
\title{retrieves a list of files from within a folder}
\usage{
getFilesFromFolder(ident, from = pwd(), filePattern = "", recurse = F)
}
\arguments{
\item{ident}{path, resource or entity ID of the folder, only one folder at a time}

\item{from}{used for relative paths, by default pwd is used, which is initiated with the step that started improveR}

\item{filePattern}{filter applied to the file name, example: *.r}

\item{recurse}{if folders within this folder are also parsed, default FALSE}
}
\description{
retrieves a list of files from within a folder
}
\references{
ics1141
}
