% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneric.R
\name{getFilesFromFolder}
\alias{getFilesFromFolder}
\title{Retrieve a List of Files from a Folder}
\usage{
getFilesFromFolder(ident, from = pwd(), filePattern = "", recurse = F)
}
\arguments{
\item{ident}{Path, resource, or entity ID of the folder (one folder at a time).}

\item{from}{Used for relative paths. By default, pwd is used (initiated with the step that started improveR).}

\item{filePattern}{Filter applied to the file name (example: \verb{*.r}).}

\item{recurse}{If TRUE, nested folders are also parsed. Defaults to FALSE.}
}
\description{
Retrieve a List of Files from a Folder
}
\references{
ics1141
}
