% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFullFolderAuditTrail.R
\name{getFullFolderAuditTrail}
\alias{getFullFolderAuditTrail}
\title{retrieves recursive audit trail for a specific folder}
\usage{
getFullFolderAuditTrail(
  ident,
  fromForRelativePathes = pwd(),
  includeReadAccess = FALSE
)
}
\arguments{
\item{ident}{the folder the audit trail is retrieved for}

\item{fromForRelativePathes}{startpoint for relative pathes, per default: pwd is used}

\item{includeReadAccess}{if TRUE also read access entries of the audit trail are shown}
}
\description{
retrieves recursive audit trail for a specific folder
}
