% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHTML.R
\name{getHTML}
\alias{getHTML}
\title{gets a HTML object}
\usage{
getHTML(ident, from = pwd(), addAsLink = TRUE, caption = "")
}
\arguments{
\item{ident}{path, resource or entity ID of the HTML file}

\item{from}{used for relative paths, by default pwd is used, which is initiated with the step that started improveR}

\item{addAsLink}{creates a link in inventory if TRUE, warn improveClean has to be run at the end}

\item{caption}{by default entityID and lastmodified are the caption, here alternative text can be provided}
}
\description{
gets a HTML object
}
\references{
ics1141
}
