% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveDirectories.R
\name{getImproveInternalDir}
\alias{getImproveInternalDir}
\title{Get improve internal directory}
\usage{
getImproveInternalDir()
}
\value{
Character string with path to internal directory
}
\description{
Gets the internal working directory for improve system files.
Provides sensible cross-platform defaults if internalFolder is not set.
}
