% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveDirectories.R
\name{getImproveWorkspaceDir}
\alias{getImproveWorkspaceDir}
\title{Get improve workspace directory}
\usage{
getImproveWorkspaceDir()
}
\value{
Character string with path to workspace directory
}
\description{
Gets the workspace directory for user files and projects.
}
