% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProcess.R
\name{getMainProcess}
\alias{getMainProcess}
\title{getMainProcess}
\usage{
getMainProcess(stepIdent)
}
\arguments{
\item{stepIdent}{resourceId of the step}
}
\description{
getMainProcess
}
\references{
ics1218
}
