% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaData.R
\name{getMetaData}
\alias{getMetaData}
\title{Gets a Metadata Data Frame}
\usage{
getMetaData(ident, from = pwd())
}
\arguments{
\item{ident}{path, resource or entity ID of the picture}

\item{from}{used for relative paths, by default pwd is used, which is initiated with the step that started improveR}
}
\description{
Gets a Metadata Data Frame
}
\references{
ics1141
}
