% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaData.R
\name{getMetaDataMap}
\alias{getMetaDataMap}
\title{loads the metadata of a given resource and writes it as string values into a list}
\usage{
getMetaDataMap(ident, from = pwd())
}
\arguments{
\item{ident}{path, resource or entity ID of the picture}

\item{from}{used for relative paths, by default pwd is used, which is initiated with the step that started improveR}
}
\description{
loads the metadata of a given resource and writes it as string values into a list
}
