% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaData.R
\name{getMetaDataMap}
\alias{getMetaDataMap}
\title{Loads the Metadata of a Given Resource and Writes It as String Values Into a List}
\usage{
getMetaDataMap(ident, from = pwd())
}
\arguments{
\item{ident}{path, resource or entity ID of the picture}

\item{from}{used for relative paths, by default pwd is used, which is initiated with the step that started improveR}
}
\description{
Loads the Metadata of a Given Resource and Writes It as String Values Into a List
}
