% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParent.R
\name{getParent}
\alias{getParent}
\title{Returns the Parent of Anything Resolved by GetCorrectId}
\usage{
getParent(identifier)
}
\arguments{
\item{identifier, }{ID or resource}
}
\description{
Returns the Parent of Anything Resolved by GetCorrectId
}
\references{
ics1088
}
