% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{getProcessFileVariables}
\alias{getProcessFileVariables}
\title{getProcessFileVariables}
\usage{
getProcessFileVariables(ident, processId)
}
\arguments{
\item{ident}{id of the step the variables refer to}

\item{processId}{id of the process the variables refer to}
}
\description{
getProcessFileVariables
}
