% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveConnect.R
\name{getRepositoryVersion}
\alias{getRepositoryVersion}
\title{Get Repository Version}
\usage{
getRepositoryVersion()
}
\value{
Character string with the repository version, or NULL if not available
}
\description{
Returns the version of the connected repository, if available.
}
