% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{getReviewComments}
\alias{getReviewComments}
\title{getReviewComments}
\usage{
getReviewComments(ident, fromForRelativePathes = pwd())
}
\arguments{
\item{ident}{id}

\item{fromForRelativePathes}{defaults to pwd, used to resolve relative pathes}
}
\description{
getReviewComments
}
\references{
ics1208
}
