% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{getReviewComments}
\alias{getReviewComments}
\title{Get Review Comments}
\usage{
getReviewComments(ident, from = pwd())
}
\arguments{
\item{ident}{id}

\item{from}{defaults to pwd, used to resolve relative paths}
}
\description{
Get Review Comments
}
\references{
ics1208
}
