% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{getReviewEntries}
\alias{getReviewEntries}
\title{Get Review Entries}
\usage{
getReviewEntries(ident, from = pwd())
}
\arguments{
\item{ident}{id}

\item{from}{defaults to pwd, used to resolve relative paths}
}
\description{
Get Review Entries
}
\references{
ics1208
}
