% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{getReviewEntryComments}
\alias{getReviewEntryComments}
\title{getReviewEtnryComments}
\usage{
getReviewEntryComments(resourceId, entryId, fromForRelativePathes = pwd())
}
\arguments{
\item{resourceId}{id (UUID) of the resource}

\item{entryId}{id (UUID) of the review entry}

\item{fromForRelativePathes}{defaults to pwd, used to resolve relative pathes}
}
\description{
getReviewEtnryComments
}
\references{
ics1543
}
