% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{getReviewers}
\alias{getReviewers}
\title{getReviewers}
\usage{
getReviewers(ident, fromForRelativePathes = pwd())
}
\arguments{
\item{ident}{id}

\item{fromForRelativePathes}{defaults to pwd, used to resolve relative pathes}
}
\description{
getReviewers
}
\references{
ics1208
}
