% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStep.R
\name{getStep}
\alias{getStep}
\title{getStep
reads a step and its processes and generates a new handle}
\usage{
getStep(ident, workflow = NULL)
}
\arguments{
\item{ident}{the ident of the step}
}
\description{
getStep
reads a step and its processes and generates a new handle
}
\references{
ics1213
}
