% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStep.R
\name{getStepInventory}
\alias{getStepInventory}
\title{getStepInventory retrieves all files from the inventory of a handle, if a step was created with this handle}
\usage{
getStepInventory(handle, recurse = F, update = T)
}
\arguments{
\item{handle}{check if a step with the same configuration already exists in the tree}

\item{recurse}{if the complete inventory should be retrieved or only the top level}

\item{update}{unloads the cached resources, default true}
}
\description{
getStepInventory retrieves all files from the inventory of a handle, if a step was created with this handle
}
\references{
ics1221
}
