% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStepEnv.R, R/runStep.R
\name{getStepResource}
\alias{getStepResource}
\title{getStepResource}
\usage{
getStepResource(handle)

getStepResource(handle)
}
\arguments{
\item{handle}{load the step resource if the handle has already been executed}
}
\description{
getStepResource

getStepResource
}
\references{
ics1221

ics1221
}
