% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStep.R
\name{getStepState}
\alias{getStepState}
\title{getStepState}
\usage{
getStepState(handle)
}
\arguments{
\item{handle}{get current state of step without caching}
}
\description{
getStepState
}
\references{
ics1221
}
