% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStep.R
\name{getStepTemplate}
\alias{getStepTemplate}
\title{getStepTemplate
reads a step and its processes and generates a new handle}
\usage{
getStepTemplate(ident)
}
\arguments{
\item{ident}{the ident of the step}
}
\description{
getStepTemplate
reads a step and its processes and generates a new handle
}
\references{
ics1213
}
