% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenRefresherPlugin.R
\name{getTokenRefresher}
\alias{getTokenRefresher}
\title{Get registered token refresher}
\usage{
getTokenRefresher(name)
}
\arguments{
\item{name}{Character string. Name of the refresher}
}
\value{
Token refresher object or NULL if not found
}
\description{
Gets a registered token refresher by name
}
