% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToolInstances.R
\name{getToolInstances}
\alias{getToolInstances}
\title{getToolInstances
returns an environement with all tool instances. in order to relaod tools from the server use resetToolInstances}
\usage{
getToolInstances()
}
\description{
getToolInstances
returns an environement with all tool instances. in order to relaod tools from the server use resetToolInstances
}
