% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleFromStep.R
\name{handleFromStep}
\alias{handleFromStep}
\title{handleFromStep
reads a step and its processes and generates a new handle}
\usage{
handleFromStep(stepId)
}
\arguments{
\item{stepId}{the ident of the step}
}
\description{
handleFromStep
reads a step and its processes and generates a new handle
}
\references{
ics1213
}
