% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importWorkflow.R
\name{importWorkflow}
\alias{importWorkflow}
\title{Import a Workflow from a Zip File into a Repository Folder}
\usage{
importWorkflow(workflowFile, importRepoFolder)
}
\arguments{
\item{workflowFile}{Character. Path to the workflow zip file to import.}

\item{importRepoFolder}{Character. Path to the repository folder where the workflow will be imported.}
}
\value{
None. The function is called for its side effects.
}
\description{
This function imports a workflow from a specified zip file into a given repository folder.
}
\examples{
\dontrun{
importWorkflow("path/to/workflow.zip", "path/to/repo/folder")
}

}
