% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerCloseFunction.R
\docType{data}
\name{improveCloseToken}
\alias{improveCloseToken}
\title{lock object}
\format{
An object of class \code{environment} of length 0.
}
\usage{
improveCloseToken
}
\description{
lock object
}
\keyword{datasets}
