% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveConnect.R
\name{improveConnect}
\alias{improveConnect}
\title{improveConnect looks for connection information and sets this connection information as default
there are three ways to set connection information
the priority is command line, environment variables and a conf.json file
\itemize{
\item command line arguments, have to be in the correct order
** token
** ID of the step that was used to initiate the connection
** the repo URL in the format https://repoaddress:repoPort/repository
\item environment variables
**    IMPROVER_TOKEN mandatory if user and password are empty
**    IMPROVER_STEP
**    IMPROVER_REPO_URL mandatory
**    IMPROVER_USER
**    IMPROVER_PASSWORD
**    IMPROVER_WORKSPACE
**    IMPROVER_SECURITY=insecure if there is no valid certificate
\item json file in the wd called conf.json
** reqToken mandatory if user and password are empty
** stepId
** repoUrl mandatory
** runWorkspace
** user
** password
}}
\usage{
improveConnect(
  logLevel = "INFO",
  secure = TRUE,
  offlinePossible = F,
  persistentCaching = F
)
}
\arguments{
\item{logLevel}{possible LogLevels: DEBUG, INFO, WARN, ERROR}

\item{secure}{if TRUE the certificates are checked}

\item{offlinePossible}{if TRUE the setup continues even if no connection is possible}

\item{persistentCaching}{default to false,
persists and reloads the caches on the filesystem in .improver.cache
if the environment variable improver.logfile is set, the logging is additionally added to this file}
}
\description{
improveConnect looks for connection information and sets this connection information as default
there are three ways to set connection information
the priority is command line, environment variables and a conf.json file
\itemize{
\item command line arguments, have to be in the correct order
** token
** ID of the step that was used to initiate the connection
** the repo URL in the format https://repoaddress:repoPort/repository
\item environment variables
**    IMPROVER_TOKEN mandatory if user and password are empty
**    IMPROVER_STEP
**    IMPROVER_REPO_URL mandatory
**    IMPROVER_USER
**    IMPROVER_PASSWORD
**    IMPROVER_WORKSPACE
**    IMPROVER_SECURITY=insecure if there is no valid certificate
\item json file in the wd called conf.json
** reqToken mandatory if user and password are empty
** stepId
** repoUrl mandatory
** runWorkspace
** user
** password
}
}
\references{
ics1081
}
