% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveConnect.R
\name{improveConnected}
\alias{improveConnected}
\title{Improve Connected}
\usage{
improveConnected(silent = FALSE)
}
\arguments{
\item{silent}{if TRUE no log message is printed}
}
\description{
improveConnected checks if improveConnect was called.
}
\seealso{
\code{\link[=improveConnect]{improveConnect()}}, \code{\link[=improveDisconnect]{improveDisconnect()}}
}
