% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveConnect.R
\name{improveConnected}
\alias{improveConnected}
\title{improveConnected checks if improveConnect was called}
\usage{
improveConnected()
}
\description{
improveConnected checks if improveConnect was called
}
\references{
ics1081
}
