% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveConnect.R
\name{improveDisconnect}
\alias{improveDisconnect}
\title{improveDisconnect removes all connection information}
\usage{
improveDisconnect()
}
\description{
improveDisconnect removes all connection information
}
\references{
ics1081
}
