% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getR.R
\name{improveInit}
\alias{improveInit}
\title{initialises modules, also automatically connects}
\usage{
improveInit(ident, from = pwd(), logLevel = "INFO", secure = T)
}
\arguments{
\item{ident}{path, resource or entity ID of the R object}

\item{from}{used for relative paths, by default from is used, which is initiated with the step that started improveR}

\item{logLevel}{logLevel for improve connect}

\item{secure}{if certificates should be checked}
}
\description{
initialises modules, also automatically connects
}
