% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauthLogin.R
\name{improveOAuth}
\alias{improveOAuth}
\title{Improve OAuth}
\usage{
improveOAuth(
  repo,
  shortEntityId = "/",
  logLevel = "INFO",
  secure = T,
  openBrowser = T,
  withCodeVerifier = T
)
}
\arguments{
\item{repo}{the repository URL in this form https://\if{html}{\out{<url>}}:\if{html}{\out{<port>}}/\if{html}{\out{<repositoryPath>}} the api part (/api/v1) is added automatically}

\item{shortEntityId}{one valid short entity ID must be provided, this is used as pathWorkingDirectory}

\item{logLevel}{possible LogLevels: DEBUG, INFO, WARN, ERROR}

\item{secure}{if TRUE the certificates are checked.
Default is TRUE, it can be set to false also by the environment variable IMPROVER_SECURITY=insecure}

\item{openBrowser}{this indicates whether the R session can open a browser for user access. Default is TRUE.
If set to FALSE, a complete verification URL is logged to the console for manual access. This enables headless authentication
for CI/CD environments or remote sessions. Can be controlled by environment variable IMPROVER_HEADLESS_OAUTH (any non-empty value).}

\item{withCodeVerifier}{if the oauth provider uses pkca code challenge verification}
}
\description{
improveOAuth is used to connect to the repository via oauth.
}
\references{
ics1081
}
