% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphics.R
\name{includeGraphics}
\alias{includeGraphics}
\title{Includes a picture in rMarkdown, with knitr::includeGraphics, used within a code block, only for a single image}
\usage{
includeGraphics(
  ident,
  from = pwd(),
  addAsLink = TRUE,
  caption = "",
  description = "",
  includeCaption = T,
  includeDescription = T,
  ...
)
}
\arguments{
\item{ident}{path, resource or entity ID of the picture}

\item{from}{used for relative paths, by default pwd is used, which is initiated with the step that started improveR}

\item{addAsLink}{creates a link in inventory if TRUE, warn improveClean has to be run at the end}

\item{caption}{by default entityID and lastmodified are the caption, here alternative text can be provided}

\item{description}{by default filename is the caption, here alternative text can be provided}

\item{includeCaption}{display the caption}

\item{includeDescription}{display the description}

\item{...}{additional options for knitr::includeGraphics}
}
\description{
Includes a picture in rMarkdown, with knitr::includeGraphics, used within a code block, only for a single image
}
\references{
ics1141
}
