% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFile.R
\name{isFileUp2Date}
\alias{isFileUp2Date}
\title{isFileUp2Date}
\usage{
isFileUp2Date(ident, from = pwd())
}
\arguments{
\item{from}{Used if a relative path is used.}
}
\description{
Checks if a new version of the file exists in the repository.
}
\references{
ics1099
}
