% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFile.R
\name{isFileUp2Date}
\alias{isFileUp2Date}
\title{isFileUp2Date}
\usage{
isFileUp2Date(ident, fromForRelativePathes = pwd())
}
\arguments{
\item{ident}{id}

\item{fromForRelativePathes}{used if a relative path is used}
}
\description{
isFileUp2Date
}
\references{
ics1099
}
