% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResource.R
\name{isResourceUp2Date}
\alias{isResourceUp2Date}
\title{isResourceUp2Date}
\usage{
isResourceUp2Date(ident, fromForRelativePathes = pwd())
}
\arguments{
\item{ident}{id}

\item{fromForRelativePathes}{pwd for relative path}
}
\description{
Checks if resource is up to date.
}
