% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerCloseFunction.R
\name{listCloseFunctions}
\alias{listCloseFunctions}
\title{lists the name of all functions registered by registerCloseFunction}
\usage{
listCloseFunctions()
}
\description{
lists the name of all functions registered by registerCloseFunction
}
