% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenRefresherPlugin.R
\name{listTokenRefreshers}
\alias{listTokenRefreshers}
\title{List registered token refreshers}
\usage{
listTokenRefreshers()
}
\value{
Character vector of refresher names
}
\description{
Lists all registered token refresher names
}
