% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAuditTrail.R
\name{loadAuditTrail}
\alias{loadAuditTrail}
\title{loads the auditTrail  by the resourceId,  entity ID or entity version ID
it uses caching
the results are returned as a data frame or a list of data frames
the dates are also converted to posix dates via convertImproveTimestampToPosix
resourceId can be a list}
\usage{
loadAuditTrail(ident, from = pwd())
}
\arguments{
\item{ident}{the resource id or the entity id of the resource}

\item{from}{used if a relative path is used}
}
\description{
arguments:
}
\references{
ics1097
}
