% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadChildResources.R
\name{loadChildResources}
\alias{loadChildResources}
\title{loads all child resources  by the resourceId,  entity ID or entity version ID
it uses caching
the results are returned as a data frame or a list of data frames
the dates are also converted to posix dates via convertImproveTimestampToPosix
resourceId can be a list}
\usage{
loadChildResources(ident, fromForRelativePathes = pwd())
}
\arguments{
\item{ident}{the resource id or the entity id of the resource}

\item{fromForRelativePathes}{used if a relative path is used}
}
\description{
arguments:
}
\references{
ics1085
}
