% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFile.R
\name{loadFile}
\alias{loadFile}
\title{loads the file  by the resourceId,  entity ID or entity version ID
it uses caching
the results are returned as a data frame or a list of data frames
the dates are also converted to posix dates via convertImproveTimestampToPosix
resourceId can be a list}
\usage{
loadFile(
  ident,
  fromForRelativePathes = pwd(),
  filePath = ".",
  addIdToName = FALSE,
  linkInInventory = FALSE
)
}
\arguments{
\item{ident}{the resource id or the entity id of the resource}

\item{fromForRelativePathes}{used if a relative path is used}

\item{filePath}{local Path where the file should be stored, relative to rootPath, normally wd}

\item{addIdToName}{boolean, if the entityId should be added to the filename}

\item{linkInInventory}{boolean, if TRUE a link to the resource is created in the inventory}
}
\description{
arguments:
}
\references{
ics1099
}
