% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFile.R
\name{loadFile}
\alias{loadFile}
\title{loadFile}
\usage{
loadFile(
  ident,
  from = pwd(),
  filePath = ".",
  addIdToName = FALSE,
  linkInInventory = FALSE
)
}
\arguments{
\item{ident}{the resource id or the entity id of the resource}

\item{from}{used if a relative path is used}

\item{filePath}{local Path where the file should be stored, relative to rootPath, normally wd}

\item{addIdToName}{logical, if the entityId should be added to the filename}

\item{linkInInventory}{logical, if TRUE a link to the resource is created in the inventory}
}
\description{
Loads a file by its resourceId, entity ID, or entity version ID.
Uses caching. The results are returned as a data frame or a list of data frames.
The dates are converted to POSIX dates with the convertImproveTimestampToPosix function.
resourceId can be a list.
}
\references{
ics1099
}
\seealso{
\code{\link[=convertImproveTimestampToPosix]{convertImproveTimestampToPosix()}}
}
