% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{loadGroup}
\alias{loadGroup}
\title{Get Group Details}
\usage{
loadGroup(groupId)
}
\arguments{
\item{groupId}{Character. The ID of the group.}
}
\value{
A list with group details.
Returns \code{NULL} if the group cannot be found.
}
\description{
Retrieves details for a specific group by its ID.
}
\examples{
\dontrun{
group <- loadGroup("some-group-id")
}
}
\seealso{
\code{\link{loadGroups}}, \code{\link{loadGroupUsers}}
}
