% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{loadGroups}
\alias{loadGroups}
\title{List All Groups}
\usage{
loadGroups()
}
\value{
A data frame of groups.
Returns \code{NULL} if the groups cannot be retrieved.
}
\description{
Retrieves a list of all groups in the improve repository.
}
\examples{
\dontrun{
groups <- loadGroups()
}
}
\seealso{
\code{\link{loadGroup}}, \code{\link{loadGroupUsers}}
}
