% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{loadGroupUsers}
\alias{loadGroupUsers}
\title{Get Group Members}
\usage{
loadGroupUsers(groupId)
}
\arguments{
\item{groupId}{Character. The ID of the group.}
}
\value{
A data frame of users in the group.
Returns \code{NULL} if the group cannot be found.
Returns \code{NULL} if the group has no members.
}
\description{
Retrieves all users belonging to a specific group.
}
\examples{
\dontrun{
members <- loadGroupUsers("some-group-id")
}
}
\seealso{
\code{\link{loadGroup}}, \code{\link{loadGroups}}
}
