% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetadata.R
\name{loadMetaData}
\alias{loadMetaData}
\title{Loads the History by the ResourceId, Entity ID or Entity Version ID
it uses caching
the results are returned as a data frame or a list of data frames
the dates are also converted to posix dates via convertImproveTimestampToPosix
resourceId can be a list}
\usage{
loadMetaData(ident, from = pwd())
}
\arguments{
\item{ident}{the resource id or the entity id of the resource}

\item{from}{used if a relative path is used}
}
\description{
arguments:
}
\references{
ics1096
}
