% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetaDataDefinitions.R
\name{loadMetaDataDefinitionPickList}
\alias{loadMetaDataDefinitionPickList}
\title{Loads One Meta Data Definition Picklist Values By Name For A Scope, The Default Scope Is "Improve Client"}
\usage{
loadMetaDataDefinitionPickList(name, scope = "Improve Client")
}
\arguments{
\item{name, }{name of the meta data definition}

\item{scope}{the metadata scope}
}
\description{
Loads One Meta Data Definition Picklist Values By Name For A Scope, The Default Scope Is "Improve Client"
}
\references{
ics1137
}
