% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetaDataDefinitions.R
\name{loadMetaDataDefinitionPickList}
\alias{loadMetaDataDefinitionPickList}
\title{loads one meta data definition picklist values by name for a scope, the default scope is "Improve Client"}
\usage{
loadMetaDataDefinitionPickList(name, scope = "Improve Client")
}
\arguments{
\item{name, }{name of the meta data definition}

\item{scope}{the metadata scope}
}
\description{
loads one meta data definition picklist values by name for a scope, the default scope is "Improve Client"
}
\references{
ics1137
}
