% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetaDataDefinitions.R
\name{loadMetaDataDefinitions}
\alias{loadMetaDataDefinitions}
\title{loads all meta data definitions for a scope, the default scope is "Improve Client"}
\usage{
loadMetaDataDefinitions(scope = "Improve Client")
}
\arguments{
\item{scope}{the metadata scope}
}
\description{
loads all meta data definitions for a scope, the default scope is "Improve Client"
}
\references{
ics1137
}
