% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadparentStep.R
\name{loadParentStep}
\alias{loadParentStep}
\title{loads the parent step of one step, not applicable to multiple steps}
\usage{
loadParentStep(ident, from = pwd())
}
\arguments{
\item{ident}{resourceID, entityId or path to the step.}

\item{from}{path working directory, default is the calling step}
}
\value{
parent step as DF
}
\description{
loads the parent step of one step, not applicable to multiple steps
}
\references{
ics1209
}
