% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadparentStep.R
\name{loadParentStep}
\alias{loadParentStep}
\title{Loads the Parent Step of One Step, Not Applicable to Multiple Steps}
\usage{
loadParentStep(ident, from = pwd())
}
\arguments{
\item{ident}{resourceID, entityId or path to the step.}

\item{from}{path working directory, default is the calling step}
}
\value{
parent step as DF
}
\description{
Loads the Parent Step of One Step, Not Applicable to Multiple Steps
}
\references{
ics1209
}
