% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceRelations.R
\name{loadRelationTypes}
\alias{loadRelationTypes}
\title{Loads All Registered Relation Types}
\usage{
loadRelationTypes(conn)
}
\arguments{
\item{conn}{database connection}
}
\description{
Loads All Registered Relation Types
}
