% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceByPathGeneric.R
\name{loadResourceByPathGeneric}
\alias{loadResourceByPathGeneric}
\title{loadResourceByPathFromServer}
\usage{
loadResourceByPathGeneric(path, from = pwd())
}
\arguments{
\item{path}{a path}

\item{from, }{starting from}
}
\description{
loadResourceByPathFromServer
}
