% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceFromServer.r
\name{loadResourceFromServer}
\alias{loadResourceFromServer}
\title{loads resource directly from the server by the resourceId, resourceVersion ID or entity ID
the results are returned as a data frame or a list of data frames
the dates are also converted to posix dates via convertImproveTimestampToPosix
if 0 is handed over, a virtual root resource is handed back
resourceId can be a list}
\usage{
loadResourceFromServer(resourceId, invalidatesReproducibility = T)
}
\arguments{
\item{resourceId}{the resource id or the entity id of the resource}

\item{invalidatesReproducibility}{this flag may only be changed by internal functions}
}
\description{
loads resource directly from the server by the resourceId, resourceVersion ID or entity ID
the results are returned as a data frame or a list of data frames
the dates are also converted to posix dates via convertImproveTimestampToPosix
if 0 is handed over, a virtual root resource is handed back
resourceId can be a list
}
