% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceRelations.R
\name{loadResourceRelations}
\alias{loadResourceRelations}
\title{Loads All Registered Resource Relations
it uses caching
the results are returned as a data frame or a list of data frames
the dates are also converted to posix dates via convertImproveTimestampToPosix
ident can be a list}
\usage{
loadResourceRelations(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
Loads All Registered Resource Relations
it uses caching
the results are returned as a data frame or a list of data frames
the dates are also converted to posix dates via convertImproveTimestampToPosix
ident can be a list
}
\references{
ics1044
}
