% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceVersionFromServer.r
\name{loadResourceVersionFromServer}
\alias{loadResourceVersionFromServer}
\title{Load Resource Version From Server}
\usage{
loadResourceVersionFromServer(entityVersionId, invalidatesReproducibility = T)
}
\arguments{
\item{entityVersionId}{the entity version id of the resource}

\item{invalidatesReproducibility}{this flag may only be changed by internal functions.}
}
\description{
Loads a resource version directly from the server by the  entity version ID.
The results are returned as a data frame or a list of data frames.
The dates are also converted to POSIX dates via convertImproveTimestampToPosix.
entityVersionId can be a list.
}
\seealso{
\code{\link[=convertImproveTimestampToPosix]{convertImproveTimestampToPosix()}},\code{\link[=loadResource]{loadResource()}}, \code{\link[=loadResourceByPathGeneric]{loadResourceByPathGeneric()}}
}
