% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{loadReviewEntries}
\alias{loadReviewEntries}
\title{loads all the review entries for a review
it uses caching
the result is returned as a data frame}
\usage{
loadReviewEntries(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
loads all the review entries for a review
it uses caching
the result is returned as a data frame
}
\references{
ics1208
}
