% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{loadReviewEntryComments}
\alias{loadReviewEntryComments}
\title{Loads All The Comments For A Review Entry}
\usage{
loadReviewEntryComments(ident, entryId, from = pwd())
}
\arguments{
\item{ident}{Identifier of the review. Can be a path, resource ID, entity ID,
or a data frame row from \code{loadResource()}.}

\item{entryId}{Character. ID (UUID) of the review entry.}

\item{from}{Base path for resolving relative paths. Defaults to \code{pwd()}.}
}
\value{
A data frame of review entry comments, or \code{NULL} if none exist.
}
\description{
Retrieves all comments for a review entry.
}
\references{
ics1543
}
