% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{loadReviewEntryComments}
\alias{loadReviewEntryComments}
\title{Loads All The Comments For A Review Entry
it uses caching
the result is returned as a data frame}
\usage{
loadReviewEntryComments(resourceId, entryId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}

\item{entryId}{id (UUID) of the review entry}
}
\description{
Loads All The Comments For A Review Entry
it uses caching
the result is returned as a data frame
}
\references{
ics1543
}
