% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{loadReviewers}
\alias{loadReviewers}
\title{Loads All The Reviewers For A Review
it uses caching
the result is returned as a data frame}
\usage{
loadReviewers(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
Loads All The Reviewers For A Review
it uses caching
the result is returned as a data frame
}
\references{
ics1208
}
