% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{loadReviews}
\alias{loadReviews}
\title{Loads All Registered Reviews
it uses caching
the result is returned as a data frame}
\usage{
loadReviews()
}
\description{
Loads All Registered Reviews
it uses caching
the result is returned as a data frame
}
\references{
ics348
}
