% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{loadReviews}
\alias{loadReviews}
\title{Loads All Registered Reviews}
\usage{
loadReviews()
}
\value{
A data frame of reviews, or \code{NULL} if none exist.
}
\description{
Retrieves the list of all reviews from the repository. Results are cached.
}
\references{
ics348
}
