% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepConfigData.R
\name{loadRunserver}
\alias{loadRunserver}
\title{loads runserver by label}
\usage{
loadRunserver(label)
}
\arguments{
\item{label}{the label of the runserver}
}
\description{
loads runserver by label
}
\references{
ics1226
}
