% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{loadToolsForCategory}
\alias{loadToolsForCategory}
\title{loadToolsForCategory}
\usage{
loadToolsForCategory(categoryId)
}
\arguments{
\item{categoryId}{categoryId of the tools}
}
\description{
loadToolsForCategory
}
\references{
ics1230
}
