% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockUnlock.R
\name{lockResource}
\alias{lockResource}
\title{lockResource}
\usage{
lockResource(ident, from = pwd())
}
\arguments{
\item{ident}{id of resource to be locked}

\item{from}{pwd for relative pathes}
}
\description{
lockResource
}
\references{
ics1139
}
