% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.R
\name{printLogs}
\alias{printLogs}
\title{printLogs prints the strucured log list to the console}
\usage{
printLogs(reset = F)
}
\arguments{
\item{reset}{T resets the list after printing}
}
\description{
printLogs prints the strucured log list to the console
}
