% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliPush.R
\name{pushCli}
\alias{pushCli}
\title{Push CLI}
\usage{
pushCli(localPath)
}
\arguments{
\item{localPath}{The local repository path.}
}
\description{
Pushes changes to the CLI. After pushing, the resource cache is invalidated
to ensure subsequent queries reflect the updated server state.
}
