% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliPushRun.R
\name{pushRunCli}
\alias{pushRunCli}
\title{Push Run CLI}
\usage{
pushRunCli(localPath, command = "Rstudio")
}
\arguments{
\item{localPath}{The local repository path.}

\item{command}{The command to run. Default is "Rstudio".}
}
\description{
Pushes and runs a command on the CLI.
}
