% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{performs this query against the elastic search}
\usage{
query(query, fetchSize = 100, fetchOffset = 0)
}
\arguments{
\item{query}{query with the same grammar as in the UI}

\item{fetchSize}{maximum number of returned results}

\item{fetchOffset}{offset of the results, enables paging}
}
\description{
performs this query against the elastic search
}
\references{
ics1143
}
