% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStep.R
\name{realiseStep}
\alias{realiseStep}
\title{getStep}
\usage{
realiseStep(handle, force = T, run = T)
}
\arguments{
\item{handle}{check if a step with the same configuration already exists in the tree}

\item{force, }{force creates a new step even if an equivalent step already exists}

\item{run, }{automatically run the step after creation (T is overridden by the setStepBreakpoint)}
}
\description{
getStep
}
\references{
ics1140
}
