% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.R
\name{redirectLogs}
\alias{redirectLogs}
\title{redirectLogs redirects the log from the default logstream to a structured list, with different context}
\usage{
redirectLogs(redirect)
}
\arguments{
\item{redirect}{T turns on redirection F turns it off}
}
\description{
redirectLogs redirects the log from the default logstream to a structured list, with different context
}
