% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveR-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{authenticatetREST}
\alias{convertImproveTimestampToPosix}
\alias{getCorrectId}
\alias{getFullFolderAuditTrail}
\alias{getParent}
\alias{getReviewComments}
\alias{getReviewEntries}
\alias{getReviewers}
\alias{improveClose}
\alias{improveConnect}
\alias{isFileUp2Date}
\alias{isResourceUp2Date}
\alias{listCloseFunctions}
\alias{loadAuditTrail}
\alias{loadChildResources}
\alias{loadFile}
\alias{loadFullChildResources}
\alias{loadHistory}
\alias{loadMetaData}
\alias{loadReferences}
\alias{loadResource}
\alias{loadResourceByPathFromServer}
\alias{loadResourceByPathGeneric}
\alias{loadResourceFromServer}
\alias{loadResourceVersionFromServer}
\alias{normalisePath}
\alias{pwd}
\alias{query}
\alias{queryFolder}
\alias{registerCloseFunction}
\alias{resetCache}
\alias{strip}
\alias{unloadAuditTrail}
\alias{unloadChildResources}
\alias{unloadFile}
\alias{unloadFullChildResources}
\alias{unloadHistory}
\alias{unloadMetaData}
\alias{unloadReferences}
\alias{unloadResource}
\alias{unregisterCloseFunction}
\alias{updateAuditTrail}
\alias{updateChildResources}
\alias{updateFile}
\alias{updateFullChildResources}
\alias{updateHistory}
\alias{updateMetaData}
\alias{updateReferences}
\alias{updateResource}
\alias{buildCaption}
\alias{buildDescriptor}
\alias{getAbstract}
\alias{getCopy}
\alias{getData}
\alias{getDataByResource}
\alias{getDesc}
\alias{getFile}
\alias{getFilesFromFolder}
\alias{getGraphics}
\alias{getHTML}
\alias{getMetaData}
\alias{getMetaDataMap}
\alias{getR}
\alias{getTextByResource}
\alias{getTextString}
\alias{includeGraphics}
\alias{resetStep}
\alias{showGraphics}
\alias{showHTML}
\alias{sourceR}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{improveRcore}{\code{\link[improveRcore]{authenticatetREST}}, \code{\link[improveRcore]{convertImproveTimestampToPosix}}, \code{\link[improveRcore]{getCorrectId}}, \code{\link[improveRcore]{getFullFolderAuditTrail}}, \code{\link[improveRcore]{getParent}}, \code{\link[improveRcore]{getReviewComments}}, \code{\link[improveRcore]{getReviewEntries}}, \code{\link[improveRcore]{getReviewers}}, \code{\link[improveRcore]{improveClose}}, \code{\link[improveRcore]{improveConnect}}, \code{\link[improveRcore]{isFileUp2Date}}, \code{\link[improveRcore]{isResourceUp2Date}}, \code{\link[improveRcore]{listCloseFunctions}}, \code{\link[improveRcore]{loadAuditTrail}}, \code{\link[improveRcore]{loadChildResources}}, \code{\link[improveRcore]{loadFile}}, \code{\link[improveRcore]{loadFullChildResources}}, \code{\link[improveRcore]{loadHistory}}, \code{\link[improveRcore]{loadMetaData}}, \code{\link[improveRcore]{loadReferences}}, \code{\link[improveRcore]{loadResource}}, \code{\link[improveRcore]{loadResourceByPathFromServer}}, \code{\link[improveRcore]{loadResourceByPathGeneric}}, \code{\link[improveRcore]{loadResourceFromServer}}, \code{\link[improveRcore]{loadResourceVersionFromServer}}, \code{\link[improveRcore]{normalisePath}}, \code{\link[improveRcore]{pwd}}, \code{\link[improveRcore]{query}}, \code{\link[improveRcore]{queryFolder}}, \code{\link[improveRcore]{registerCloseFunction}}, \code{\link[improveRcore]{resetCache}}, \code{\link[improveRcore]{strip}}, \code{\link[improveRcore]{unloadAuditTrail}}, \code{\link[improveRcore]{unloadChildResources}}, \code{\link[improveRcore]{unloadFile}}, \code{\link[improveRcore]{unloadFullChildResources}}, \code{\link[improveRcore]{unloadHistory}}, \code{\link[improveRcore]{unloadMetaData}}, \code{\link[improveRcore]{unloadReferences}}, \code{\link[improveRcore]{unloadResource}}, \code{\link[improveRcore]{unregisterCloseFunction}}, \code{\link[improveRcore]{updateAuditTrail}}, \code{\link[improveRcore]{updateChildResources}}, \code{\link[improveRcore]{updateFile}}, \code{\link[improveRcore]{updateFullChildResources}}, \code{\link[improveRcore]{updateHistory}}, \code{\link[improveRcore]{updateMetaData}}, \code{\link[improveRcore]{updateReferences}}, \code{\link[improveRcore]{updateResource}}}

  \item{improveRget}{\code{\link[improveRget]{buildCaption}}, \code{\link[improveRget]{buildDescriptor}}, \code{\link[improveRget]{getAbstract}}, \code{\link[improveRget]{getCopy}}, \code{\link[improveRget]{getData}}, \code{\link[improveRget]{getDataByResource}}, \code{\link[improveRget]{getDesc}}, \code{\link[improveRget]{getFile}}, \code{\link[improveRget]{getFilesFromFolder}}, \code{\link[improveRget]{getGraphics}}, \code{\link[improveRget]{getHTML}}, \code{\link[improveRget]{getMetaData}}, \code{\link[improveRget]{getMetaDataMap}}, \code{\link[improveRget]{getR}}, \code{\link[improveRget]{getTextByResource}}, \code{\link[improveRget]{getTextString}}, \code{\link[improveRget]{includeGraphics}}, \code{\link[improveRget]{resetStep}}, \code{\link[improveRget]{showGraphics}}, \code{\link[improveRget]{showHTML}}, \code{\link[improveRget]{sourceR}}}
}}

