% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFullChildResources.R
\name{refreshFullChildResources}
\alias{refreshFullChildResources}
\alias{updateFullChildResources}
\title{refreshFullChildResources}
\usage{
refreshFullChildResources(ident)

updateFullChildResources(...)
}
\arguments{
\item{ident}{id}
}
\description{
refreshFullChildResources
}
\references{
ics1085
}
