% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFullChildResources.R
\name{refreshFullChildResources}
\alias{refreshFullChildResources}
\alias{updateFullChildResources}
\title{refreshFullChildResources}
\usage{
refreshFullChildResources(...)


updateFullChildResources(ident)
}
\arguments{
\item{ident}{id}
}
\description{
\code{refreshFullChildResources} is deprecated; use \code{refreshFullChildResources} instead.

updateFullChildResources
}
\references{
ics1085
}
