% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetadata.R
\name{refreshMetaData}
\alias{refreshMetaData}
\alias{updateMetaData}
\title{Update Meta Data}
\usage{
refreshMetaData(...)


updateMetaData(ident)
}
\arguments{
\item{ident}{id}
}
\description{
\code{refreshMetaData} is deprecated; use \code{refreshMetaData} instead.

Update Meta Data
}
\references{
ics1096
}
