% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetadata.R
\name{refreshMetaData}
\alias{refreshMetaData}
\alias{updateMetaData}
\title{Refresh Meta Data}
\usage{
refreshMetaData(ident)

updateMetaData(...)
}
\arguments{
\item{ident}{id}
}
\description{
Refresh Meta Data
}
\references{
ics1096
}
