% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetaDataDefinitions.R
\name{refreshMetaDataDefinitions}
\alias{refreshMetaDataDefinitions}
\alias{updateMetaDataDefinitions}
\title{Refresh Meta Data Definitions}
\usage{
refreshMetaDataDefinitions(scope = "Improve Client")

updateMetaDataDefinitions(...)
}
\arguments{
\item{scope}{the metadata scope}

\item{...}{For backwards compatibility with the deprecated \verb{update*} alias; not used by \verb{refresh*} itself.}
}
\description{
Refresh Meta Data Definitions
}
\references{
ics1137
}
