% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadparentStep.R
\name{refreshParentStep}
\alias{refreshParentStep}
\alias{updateParentStep}
\title{Refresh Parent Step from Server}
\usage{
refreshParentStep(ident, from = pwd())

updateParentStep(...)
}
\arguments{
\item{ident}{resourceID, entityId or path to the step.}

\item{from}{path working directory, default is the calling step}
}
\description{
Refresh Parent Step from Server
}
\references{
ics1209
}
