% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReferences.R
\name{refreshReferences}
\alias{refreshReferences}
\alias{updateReferences}
\title{Refresh References}
\usage{
refreshReferences(ident)

updateReferences(...)
}
\arguments{
\item{ident}{id}
}
\description{
Refresh References
}
\references{
ics1206
}
