% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReferences.R
\name{refreshReferences}
\alias{refreshReferences}
\alias{updateReferences}
\title{Update References}
\usage{
refreshReferences(...)


updateReferences(ident)
}
\arguments{
\item{ident}{id}
}
\description{
\code{refreshReferences} is deprecated; use \code{refreshReferences} instead.

Update References
}
\references{
ics1206
}
