% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceRelations.R
\name{refreshRelationTypes}
\alias{refreshRelationTypes}
\alias{updateRelationTypes}
\title{Reloads the Relation Types}
\usage{
refreshRelationTypes()

updateRelationTypes(...)
}
\description{
Reloads the Relation Types
}
