% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResource.R
\name{refreshResource}
\alias{refreshResource}
\alias{updateResource}
\title{Refresh Resource}
\usage{
refreshResource(ident, from = pwd())

updateResource(...)
}
\arguments{
\item{ident}{id}

\item{from}{pwd for relative path}
}
\description{
Refreshes a resource from the server by clearing cache and reloading.
}
\references{
ics1090
}
