% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResource.R
\name{refreshResource}
\alias{refreshResource}
\alias{updateResource}
\title{Update Resource}
\usage{
refreshResource(...)


updateResource(ident, from = pwd())
}
\arguments{
\item{ident}{id}

\item{from}{pwd for relative path}
}
\description{
\code{refreshResource} is deprecated; use \code{refreshResource} instead.

Updates a resource.
}
\references{
ics1090
}
