% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReview.R
\name{refreshReviews}
\alias{refreshReviews}
\alias{updateReviews}
\title{Reloads the Reviews}
\usage{
refreshReviews(...)


updateReviews()
}
\value{
A data frame of reviews, or \code{NULL} if none exist.
}
\description{
\code{refreshReviews} is deprecated; use \code{refreshReviews} instead.

Reloads the Reviews
}
\references{
ics348
}
