% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{refreshToken}
\alias{refreshToken}
\title{refreshToken requests a new token to access the system. Just used in combination with improveLogin, run tokens do not need to be refreshed}
\usage{
refreshToken(alwaysRefresh = F)
}
\arguments{
\item{alwaysRefresh}{refresh no matter how much time has elapsed}
}
\description{
refreshToken requests a new token to access the system. Just used in combination with improveLogin, run tokens do not need to be refreshed
}
\references{
ics1208
}
