% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{refreshToken}
\alias{refreshToken}
\title{refreshToken}
\usage{
refreshToken(alwaysRefresh = F)
}
\arguments{
\item{alwaysRefresh}{refresh no matter how much time has elapsed}
}
\description{
refreshToken requests a new token to access the system.
Run tokens do not need to be refreshed.
}
\references{
ics1208
}
\seealso{
\code{\link[=improveLogin]{improveLogin()}}
}
