% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenRefresherPlugin.R
\name{registerTokenRefresher}
\alias{registerTokenRefresher}
\title{Register a token refresher plugin}
\usage{
registerTokenRefresher(name, refresher)
}
\arguments{
\item{name}{Character string. Name of the refresher}

\item{refresher}{List with required methods: init, start, stop, isRunning, getToken}
}
\description{
Registers a token refresher implementation
}
