% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{removeGroupUser}
\alias{removeGroupUser}
\title{Remove a User from a Group}
\usage{
removeGroupUser(groupId, userId)
}
\arguments{
\item{groupId}{Character. The ID of the group.}

\item{userId}{Character. The ID of the user to remove.}
}
\value{
\code{TRUE} if the user was removed successfully, \code{FALSE} otherwise.
}
\description{
Removes a user from an existing group.
}
\examples{
\dontrun{
removeGroupUser("group-id", "user-id")
}
}
\seealso{
\code{\link{addGroupUser}}, \code{\link{loadGroupUsers}}
}
