% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{removeStepGridArguments}
\alias{removeStepGridArguments}
\title{removes all grid arguments from the step}
\usage{
removeStepGridArguments(stepHandle)
}
\arguments{
\item{stepHandle}{id of the prepared step}
}
\description{
removes all grid arguments from the step
}
