% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromCache.R
\name{resetCache}
\alias{resetCache}
\title{empties the complete cache and removes a cacheFile if it exists
sets the step to non reproducible
same internal state like directly after calling improveConnect}
\usage{
resetCache()
}
\description{
empties the complete cache and removes a cacheFile if it exists
sets the step to non reproducible
same internal state like directly after calling improveConnect
}
\references{
ics1091
}
