% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFromCache.R
\name{resetCache}
\alias{resetCache}
\title{Reset Cache}
\usage{
resetCache()
}
\description{
Empties the entire cache and removes a cacheFile, if it exists.
Sets the step to non-reproducible. The reset cache is similar to
its state after calling improveConnect.
}
\references{
ics1091
}
\seealso{
\code{\link[=improveConnect]{improveConnect()}}
}
