% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToolInstances.R
\name{resetToolInstances}
\alias{resetToolInstances}
\title{resetToolInstances
clears the cache for toolInstances and reloads everything from the server}
\usage{
resetToolInstances()
}
\description{
resetToolInstances
clears the cache for toolInstances and reloads everything from the server
}
