% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStep.R
\name{runStepResource}
\alias{runStepResource}
\title{runStepResource}
\usage{
runStepResource(ident)
}
\arguments{
\item{ident}{check if a step with the same configuration already exists in the tree}
}
\description{
runStepResource
}
\references{
ics1140
}
