% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editable.R
\name{setEditable}
\alias{setEditable}
\title{setEditable}
\usage{
setEditable(editable = T)
}
\arguments{
\item{editable}{if true it is possible to edit contents in the repository}
}
\description{
setEditable sets the editable mode, default true
}
\seealso{
\code{\link[=improveConnected]{improveConnected()}}
}
