% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.R
\name{setLogContext}
\alias{setLogContext}
\title{setLogContext sets the context for the structured log list}
\usage{
setLogContext(context)
}
\arguments{
\item{context}{a string, the context name}
}
\description{
setLogContext sets the context for the structured log list
}
