% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{setProcessVariables}
\alias{setProcessVariables}
\title{setProcessVariables}
\usage{
setProcessVariables(
  stepId,
  processId,
  processType = "main",
  name = "Main",
  selected = TRUE,
  runserverId = "",
  toolId = "",
  runserverToolId = "",
  toolIgnorePatterns = "",
  mainProcess = T,
  gridTool,
  toolArguments = NULL,
  position = 1,
  parentProcessId = NULL,
  toolBrowserUrl = NULL,
  toolDeletePatterns = NULL,
  toolStreamablePatterns = ""
)
}
\arguments{
\item{stepId}{id of the step the variables refer to}

\item{processId}{id of the process the variables refer to}

\item{processType}{processType}

\item{name}{name}

\item{selected}{selected}

\item{runserverId}{runserverId}

\item{toolId}{toolId}

\item{runserverToolId}{runserverToolId}

\item{toolIgnorePatterns}{toolIgnorePatterns}

\item{mainProcess}{if this process is status relevant, there needs to be exactly one mainProcess}

\item{gridTool}{gridTool}

\item{toolArguments}{toolArguments}

\item{position}{the position in the process list}

\item{parentProcessId}{only to use when processType equals sub}

\item{toolBrowserUrl}{a URL improve uses to automatically open while the step is running}

\item{toolDeletePatterns}{files that wont get checked in}

\item{toolStreamablePatterns}{file that can be streamed to monitor the step}
}
\description{
setProcessVariables
}
