% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setStepCommandLine}
\alias{setStepCommandLine}
\title{sets the command line}
\usage{
setStepCommandLine(stepHandle, commandline, append = T)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{commandline}{the new command line or command line appendix}

\item{append}{if the complete command line is replaced or this is appended defaults to TRUE}
}
\description{
sets the command line
}
