% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setStepRunserverName}
\alias{setStepRunserverName}
\title{sets the runserver by name}
\usage{
setStepRunserverName(stepHandle, runserverName)
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{runserverName}{name of the runserver}
}
\description{
sets the runserver by name
}
