% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepHandleHelpers.R
\name{setStepRunserverToolName}
\alias{setStepRunserverToolName}
\title{sets the runserver tool by name}
\usage{
setStepRunserverToolName(stepHandle, runserverToolName, process = "Main")
}
\arguments{
\item{stepHandle}{id of the prepared step}

\item{runserverToolName}{name of the runserver tool}

\item{process}{the name of the process, default = Main. If the process does not yet exist it is created}
}
\description{
sets the runserver tool by name
}
