% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphics.R
\name{showGraphics}
\alias{showGraphics}
\title{Includes a picture in rMarkdown, creates a string that has to be outputted}
\usage{
showGraphics(
  ident,
  from = pwd(),
  addAsLink = TRUE,
  caption = "",
  description = "",
  includeCaption = T,
  includeDescription = T
)
}
\arguments{
\item{ident}{path, resource or entity ID of the picture}

\item{from}{used for relative paths, by default pwd is used, which is initiated with the step that started improveR}

\item{addAsLink}{creates a link in inventory if TRUE, warn improveClean has to be run at the end}

\item{caption}{by default entityID and lastmodified are the caption, here alternative text can be provided}

\item{description}{by default filename is the caption, here alternative text can be provided}

\item{includeCaption}{display the caption}

\item{includeDescription}{display the description}
}
\description{
Includes a picture in rMarkdown, creates a string that has to be outputted
}
\references{
ics1141
}
