% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadChildResources.R
\name{unloadChildResources}
\alias{unloadChildResources}
\title{unloadChildResources}
\usage{
unloadChildResources(ident)
}
\arguments{
\item{ident}{id}
}
\description{
unloadChildResources
}
\references{
ics1085
}
