% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadChildResources.R
\name{unloadChildResources}
\alias{unloadChildResources}
\title{Unload Child Resources from Cache}
\usage{
unloadChildResources(ident)
}
\arguments{
\item{ident}{The resource identifier - can be a resource ID, entity ID,
entity version ID, or path.}
}
\value{
Invisibly returns NULL. Called for side effect of clearing cache.
}
\description{
Removes child resources data from the cache for the specified resource.
}
\references{
ics1085
}
\seealso{
\code{\link{loadChildResources}} to load child resources,
\code{\link{refreshChildResources}} to refresh from server
}
