% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadChildResources.R
\name{unloadChildResources}
\alias{unloadChildResources}
\title{unloadChildResources}
\usage{
unloadChildResources(ident)
}
\description{
Removes a resource from the cache.
}
\references{
ics1085 # QUESTION # @HACKLM ics documents are not customer/r-package user facing; could all be removed -correct?
}
