% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFile.R
\name{unloadFile}
\alias{unloadFile}
\title{unloadFile}
\usage{
unloadFile(ident, from = pwd(), filePath = ".", addIdToName = FALSE)
}
\arguments{
\item{from}{used if a relative path is used}

\item{filePath}{local Path where the file should be stored, relative to rootPath, normally wd}

\item{addIdToName}{logical, if the entityId should be added to the filename}
}
\description{
Removes a file from cache.
}
\references{
ics1099
}
