% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFile.R
\name{unloadFile}
\alias{unloadFile}
\title{unloadFile}
\usage{
unloadFile(ident, from = pwd(), filePath = ".", addIdToName = FALSE)
}
\arguments{
\item{ident}{id}

\item{from}{used if a relative path is used}

\item{filePath}{local Path where the file should be stored, relative to rootPath, normally wd}

\item{addIdToName}{logical, if the entityId should be added to the filename}
}
\description{
Removes a file from cache.
}
\section{Details ident}{

There are multiple ways to describe the ident of a resource:

Absolute idents:
\itemize{
\item resourceId: a UUID
\item Data frame: uses the resourceId value of the data frame
\item entityId: pointer to the latest version of a resource. Short and long entityIds are accepted'
\item entityVersionId: pointer to a specific version of a resource. Short and long entityIds are accepted'
\item path: the full path to a resource, always starting with /.
}

Relative idents:
\itemize{
\item All relative idents are path based, they always have to start with ./ or ../'
\item Relative path without pwd: always starts from the return value of pwd()'
\item Relative path and pwd as second argument: starts the relative path from the
absolute ident that was handed over as second argument. Sometimes still called
“from” but will be updated to pwd.
}
}

\references{
ics1099
}
