% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFile.R
\name{unloadFile}
\alias{unloadFile}
\title{unloadFile}
\usage{
unloadFile(
  ident,
  fromForRelativePathes = pwd(),
  filePath = ".",
  addIdToName = FALSE
)
}
\arguments{
\item{ident}{id}

\item{fromForRelativePathes}{used if a relative path is used}

\item{filePath}{local Path where the file should be stored, relative to rootPath, normally wd}

\item{addIdToName}{boolean, if the entityId should be added to the filename}
}
\description{
unloadFile
}
\references{
ics1099
}
