% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFullChildResources.R
\name{unloadFullChildResources}
\alias{unloadFullChildResources}
\title{Unload Full Child Resources}
\usage{
unloadFullChildResources(ident)
}
\arguments{
\item{ident}{id}
}
\description{
Unload Full Child Resources
}
\references{
ics1085
}
