% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetaDataDefinitions.R
\name{unloadMetaDataDefinitions}
\alias{unloadMetaDataDefinitions}
\title{Unload Meta Data Definitions}
\usage{
unloadMetaDataDefinitions(scope = "Improve Client")
}
\arguments{
\item{scope}{the metadata scope}
}
\description{
Unload Meta Data Definitions
}
\references{
ics1137
}
