% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReferences.R
\name{unloadReferences}
\alias{unloadReferences}
\title{Unload References}
\usage{
unloadReferences(ident)
}
\arguments{
\item{ident}{id}
}
\description{
Unload References
}
\references{
ics1206
}
