% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceRelations.R
\name{unloadRelationTypes}
\alias{unloadRelationTypes}
\title{unloads all relation types}
\usage{
unloadRelationTypes(conn)
}
\arguments{
\item{conn}{database connection}
}
\description{
unloads all relation types
}
