% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResource.R
\name{unloadResource}
\alias{unloadResource}
\title{unloadResource}
\usage{
unloadResource(ident, from = pwd())
}
\arguments{
\item{ident}{id}

\item{from}{pwd for relative path}
}
\description{
Unloads a resource.
}
\references{
ics1090
}
