% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceRelations.R
\name{unloadResourceRelations}
\alias{unloadResourceRelations}
\title{Unloads All Resource Relations}
\usage{
unloadResourceRelations(ident, from = pwd())
}
\arguments{
\item{ident}{Identifier of the resource.}

\item{from}{Base path for resolving relative paths. Defaults to \code{pwd()}.}
}
\description{
Unloads All Resource Relations
}
\references{
ics1044
}
