% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{unloadReviewComments}
\alias{unloadReviewComments}
\title{unloads the review comments for a review}
\usage{
unloadReviewComments(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
unloads the review comments for a review
}
\references{
ics1208
}
