% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{unloadReviewComments}
\alias{unloadReviewComments}
\title{Unloads the Review Comments for a Review}
\usage{
unloadReviewComments(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
Unloads the Review Comments for a Review
}
\references{
ics1208
}
