% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{unloadReviewEntries}
\alias{unloadReviewEntries}
\title{unloads the review entries for a review}
\usage{
unloadReviewEntries(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
unloads the review entries for a review
}
\references{
ics1208
}
