% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{unloadReviewEntries}
\alias{unloadReviewEntries}
\title{Unloads the Review Entries for a Review}
\usage{
unloadReviewEntries(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
Unloads the Review Entries for a Review
}
\references{
ics1208
}
