% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{unloadReviewEntryComments}
\alias{unloadReviewEntryComments}
\title{unloads the comments for a review entry}
\usage{
unloadReviewEntryComments(resourceId, entryId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}

\item{entryId}{id (UUID) of the review entry}
}
\description{
unloads the comments for a review entry
}
\references{
ics1543
}
