% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{unloadReviewers}
\alias{unloadReviewers}
\title{Unloads the Reviewers for a Review}
\usage{
unloadReviewers(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
Unloads the Reviewers for a Review
}
\references{
ics1208
}
