% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{unloadReviewers}
\alias{unloadReviewers}
\title{unloads the reviewers for a review}
\usage{
unloadReviewers(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
unloads the reviewers for a review
}
\references{
ics1208
}
