% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockUnlock.R
\name{unlockResource}
\alias{unlockResource}
\title{unlockResource}
\usage{
unlockResource(ident, from = pwd())
}
\arguments{
\item{ident}{id of resource to be locked}

\item{from}{pwd for relative pathes}
}
\description{
unlockResource
}
\references{
ics1139
}
