% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerCloseFunction.R
\name{unregisterCloseFunction}
\alias{unregisterCloseFunction}
\title{unregisterCloseFunction removes a close function by the name it was registered with}
\usage{
unregisterCloseFunction(name)
}
\arguments{
\item{name}{the registered name}
}
\description{
unregisterCloseFunction removes a close function by the name it was registered with
}
