% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadChildResources.R
\name{updateChildResources}
\alias{updateChildResources}
\title{updateChildResources}
\usage{
updateChildResources(ident)
}
\description{
Updates a child resource by removing it from the cache and subsequently loading
it from the server.
}
\references{
ics1085
}
\seealso{
\code{\link[=loadChildResources]{loadChildResources()}}, \code{\link[=unloadChildResources]{unloadChildResources()}}
}
