% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadChildResources.R
\name{updateChildResources}
\alias{updateChildResources}
\title{updateChildResources}
\usage{
updateChildResources(ident)
}
\arguments{
\item{ident}{id}
}
\description{
updateChildResources
}
\references{
ics1085
}
