% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFile.R
\name{updateFile}
\alias{updateFile}
\title{updateFile}
\usage{
updateFile(
  ident,
  from = pwd(),
  filePath = ".",
  addIdToName = FALSE,
  linkInInventory = FALSE
)
}
\arguments{
\item{from}{Used if a relative path is used.}

\item{filePath}{Local path where the file is stored (relative to rootPath).}

\item{addIdToName}{Logical; if TRUE the entity id is added to the filename.}

\item{linkInInventory}{Logical; if TRUE, creates a link to the resource in inventory.}
}
\description{
Retrieves5 the latest version of the file from the repository.
}
\references{
ics1099
}
