% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFile.R
\name{updateFile}
\alias{updateFile}
\title{updateFile}
\usage{
updateFile(
  ident,
  fromForRelativePathes = pwd(),
  filePath = ".",
  addIdToName = FALSE,
  linkInInventory = FALSE
)
}
\arguments{
\item{ident}{id}

\item{fromForRelativePathes}{used if a relative path is used}

\item{filePath}{local Path where the file should be stored, relative to rootPath, normally wd}

\item{addIdToName}{boolean, if the entityId should be added to the filename}

\item{linkInInventory}{boolean, if TRUE a link to the resource is created in the inventory}
}
\description{
updateFile
}
\references{
ics1099
}
