% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{updateFileContent}
\alias{updateFileContent}
\title{updates file content,
prerequisite, must be a file
multiple files can be updated at once}
\usage{
updateFileContent(ident, localPath, comment = "modified by improveRW")
}
\arguments{
\item{ident}{the resources to be updated}

\item{localPath}{the path to the file with the new fileContent}

\item{comment}{comment for the commit, defaults to "modified by improveRW"}
}
\description{
updates file content,
prerequisite, must be a file
multiple files can be updated at once
}
\references{
ics1210
}
