% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{updateFileContent}
\alias{updateFileContent}
\title{Update File Content
Prerequisite: must be a file.
Multiple files can be updated at once.}
\usage{
updateFileContent(ident, localPath, comment = "modified by improveRW")
}
\arguments{
\item{ident}{Resource(s) to be updated.}

\item{localPath}{Path to the file with the new content.}

\item{comment}{Commit comment. Defaults to "modified by improveRW".}
}
\value{
The refreshed resource after the upload, as \code{\link[=refreshResource]{refreshResource()}} returns it.
\code{NULL} when \code{ident} resolves to no resource or the upload itself failed. For several
idents the refreshed resources of all of them, merged.
}
\description{
Update File Content
Prerequisite: must be a file.
Multiple files can be updated at once.
}
\references{
ics1210
}
