% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFullChildResources.R
\name{updateFullChildResources}
\alias{updateFullChildResources}
\title{updateFullChildResources}
\usage{
updateFullChildResources(ident)
}
\arguments{
\item{ident}{id}
}
\description{
updateFullChildResources
}
\references{
ics1085
}
