% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetaDataDefinitions.R
\name{updateMetaDataDefinitions}
\alias{updateMetaDataDefinitions}
\title{Update Meta Data Definitions}
\usage{
updateMetaDataDefinitions(scope = "Improve Client")
}
\arguments{
\item{scope}{the metadata scope}
}
\description{
Update Meta Data Definitions
}
\references{
ics1137
}
