% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{updateMetaDate}
\alias{updateMetaDate}
\title{Updates Metadata Value To A Resource}
\usage{
updateMetaDate(ident, descriptorName, value)
}
\arguments{
\item{ident}{the resource this metadata value is attached to}

\item{descriptorName}{name of the metadate}

\item{value}{the value has to match the type of the descriptor. For dates, posix date has to be used, for LOVs the text has to exist}
}
\description{
Updates Metadata Value To A Resource
}
\references{
ics1137
}
