% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadparentStep.R
\name{updateParentStep}
\alias{updateParentStep}
\title{UpdateParentStep Reloads The Runservers From The Repository}
\usage{
updateParentStep(ident, from = pwd())
}
\arguments{
\item{ident}{resourceID, entityId or path to the step.}

\item{from}{path working directory, default is the calling step}
}
\description{
UpdateParentStep Reloads The Runservers From The Repository
}
\references{
ics1209
}
