% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceRelations.R
\name{updateRelationTypes}
\alias{updateRelationTypes}
\title{reloads the relation types}
\usage{
updateRelationTypes(conn)
}
\arguments{
\item{conn}{database connection}
}
\description{
reloads the relation types
}
