% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceRelations.R
\name{updateResourceRelation}
\alias{updateResourceRelation}
\title{Updates a Resource Relation}
\usage{
updateResourceRelation(
  resourceId,
  relationId,
  newRelationTypeId,
  newDescription
)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}

\item{relationId}{id (UUID) of the relation}

\item{newRelationTypeId}{id (UUID) of the relation type}

\item{newDescription}{description of the resource relation}
}
\description{
Updates a Resource Relation
}
\references{
ics1044
}
