% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResourceRelations.R
\name{updateResourceRelations}
\alias{updateResourceRelations}
\title{Reloads the Resource Relations}
\usage{
updateResourceRelations(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
Reloads the Resource Relations
}
\references{
ics1044
}
