% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{updateReviewComments}
\alias{updateReviewComments}
\title{reloads the review comments for a review}
\usage{
updateReviewComments(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
reloads the review comments for a review
}
\references{
ics1208
}
