% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{updateReviewEntries}
\alias{updateReviewEntries}
\title{reloads the review entries for a review}
\usage{
updateReviewEntries(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
reloads the review entries for a review
}
\references{
ics1208
}
