% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{updateReviewEntries}
\alias{updateReviewEntries}
\title{Reloads the Review Entries for a Review}
\usage{
updateReviewEntries(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
Reloads the Review Entries for a Review
}
\references{
ics1208
}
