% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{updateReviewEntryComments}
\alias{updateReviewEntryComments}
\title{reloads the comments for a review entry}
\usage{
updateReviewEntryComments(resourceId, entryId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}

\item{entryId}{id (UUID) of the review entry}
}
\description{
reloads the comments for a review entry
}
\references{
ics1543
}
