% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{updateReviewEntryComments}
\alias{updateReviewEntryComments}
\title{Reloads the Comments for a Review Entry}
\usage{
updateReviewEntryComments(resourceId, entryId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}

\item{entryId}{id (UUID) of the review entry}
}
\description{
Reloads the Comments for a Review Entry
}
\references{
ics1543
}
