% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{updateReviewers}
\alias{updateReviewers}
\title{reloads the reviewers for a review}
\usage{
updateReviewers(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
reloads the reviewers for a review
}
\references{
ics1208
}
