% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadREview.R
\name{updateReviewers}
\alias{updateReviewers}
\title{Reloads the Reviewers for a Review}
\usage{
updateReviewers(resourceId)
}
\arguments{
\item{resourceId}{id (UUID) of the resource}
}
\description{
Reloads the Reviewers for a Review
}
\references{
ics1208
}
