% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{uploadFolder}
\alias{uploadFolder}
\title{uploads a complete folder
prerequisite, must be a folder}
\usage{
uploadFolder(targetIdent, localFolder, comment = "modified by improveRW")
}
\arguments{
\item{targetIdent}{the target resource}

\item{localFolder}{the path to the file with the new fileContent}

\item{comment}{comment for the commit, defaults to "modified by improveRW"}
}
\description{
uploads a complete folder
prerequisite, must be a folder
}
\references{
ics1210
}
