% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{uploadFolder}
\alias{uploadFolder}
\title{Upload a Complete Folder
Prerequisite: must be a folder.}
\usage{
uploadFolder(targetIdent, localFolder, comment = "modified by improveRW")
}
\arguments{
\item{targetIdent}{Target resource.}

\item{localFolder}{Path to the local folder to upload.}

\item{comment}{Commit comment. Defaults to "modified by improveRW".}
}
\description{
Upload a Complete Folder
Prerequisite: must be a folder.
}
\references{
ics1210
}
